==> src/01_2_transform_hogar.R <==
pacman::p_load(tidyverse, haven, here)

hogar <- read_rds(here("data", "01_intermediate", "hogar.rds")) %>%
  mutate(
    ## Hogar, identificación
    
    departamento_auxiliar = departamento,
    departamento = case_when(
      departamento == "01" ~ "Montevideo",
      departamento == "02" ~ "Artigas",
      departamento == "03" ~ "Canelones",
      departamento == "04" ~ "Colonia",
      departamento == "05" ~ "Durazno",
      departamento == "06" ~ "Flores",
      departamento == "07" ~ "Florida",
      departamento == "08" ~ "Lavalleja",
      departamento == "09" ~ "Maldonado",
      departamento == "10" ~ "Paysandú",
      departamento == "11" ~ "Río Negro",
      departamento == "12" ~ "Rivera",
      departamento == "13" ~ "Rocha",
      departamento == "14" ~ "Salto",
      departamento == "15" ~ "San José",
      departamento == "16" ~ "Soriano",
      departamento == "17" ~ "Tacuarembó",
      departamento == "18" ~ "Treinta y Tres",
      departamento == "19" ~ "Cerro Largo"
    ),
    localidad_auxiliar = case_when(
      departamento_auxiliar == "01" ~ paste0("01", localidad),
      departamento_auxiliar == "02" ~ paste0("02", localidad),
      departamento_auxiliar == "03" ~ paste0("03", localidad),
      departamento_auxiliar == "04" ~ paste0("04", localidad),
      departamento_auxiliar == "05" ~ paste0("05", localidad),
      departamento_auxiliar == "06" ~ paste0("06", localidad),
      departamento_auxiliar == "07" ~ paste0("07", localidad),
      departamento_auxiliar == "08" ~ paste0("08", localidad),
      departamento_auxiliar == "09" ~ paste0("09", localidad),
      departamento_auxiliar == "10" ~ paste0("10", localidad),
      departamento_auxiliar == "11" ~ paste0("11", localidad),
      departamento_auxiliar == "12" ~ paste0("12", localidad),
      departamento_auxiliar == "13" ~ paste0("13", localidad),
      departamento_auxiliar == "14" ~ paste0("14", localidad),
      departamento_auxiliar == "15" ~ paste0("15", localidad),
      departamento_auxiliar == "16" ~ paste0("16", localidad),
      departamento_auxiliar == "17" ~ paste0("17", localidad),
      departamento_auxiliar == "18" ~ paste0("18", localidad),
      departamento_auxiliar == "19" ~ paste0("19", localidad),
    ),
    localidad_nombre = case_when(
      localidad_auxiliar == "01020" ~	"MONTEVIDEO",
      localidad_auxiliar == "01621" ~	"ABAYUBÁ",
      localidad_auxiliar == "01622" ~	"SANTIAGO VÁZQUEZ",
      localidad_auxiliar == "01721" ~	"PAJAS BLANCAS",
      localidad_auxiliar == "01900" ~	"RURAL/PARAJE RURAL",
      localidad_auxiliar == "02020" ~	"ARTIGAS",
      localidad_auxiliar == "02021" ~	"BELLA UNION",
      localidad_auxiliar == "02522" ~	"TOMAS GOMENSORO",
      localidad_auxiliar == "02621" ~	"BALTASAR BRUM",
      localidad_auxiliar == "02622" ~	"ARROCERA SAN PEDRO",
      localidad_auxiliar == "02624" ~	"CAINSA CAMPO 3",
      localidad_auxiliar == "02721" ~	"BERNABE RIVERA",
      localidad_auxiliar == "02722" ~	"CALPICA ITACUMBÚ",
      localidad_auxiliar == "02723" ~	"COLONIA RIVERA",
      localidad_auxiliar == "02724" ~	"FRANQUIA",
      localidad_auxiliar == "02725" ~	"SEQUEIRA",
      localidad_auxiliar == "02821" ~	"CAINSA 2 ITACUMBÚ",
      localidad_auxiliar == "02822" ~	"CORONADO",
      localidad_auxiliar == "02823" ~	"CUAREIM",
      localidad_auxiliar == "02824" ~	"CUARO",
      localidad_auxiliar == "02825" ~	"GUAYUBIRA",
      localidad_auxiliar == "02826" ~	"JAVIER DE VIANA",
      localidad_auxiliar == "02827" ~	"LA BOLSA",
      localidad_auxiliar == "02828" ~	"LAS PIEDRAS",
      localidad_auxiliar == "02829" ~	"PAGUERO",
      localidad_auxiliar == "02830" ~	"PINTADITO",
      localidad_auxiliar == "02831" ~	"PORT. DE HIERRO Y CAMPODONICO",
      localidad_auxiliar == "02832" ~	"TAMANDUA",
      localidad_auxiliar == "02833" ~	"MONES QUINTELA",
      localidad_auxiliar == "02900" ~	"RURAL/PARAJE RURAL",
      localidad_auxiliar == "02921" ~	"ARROCERA RIUSA",
      localidad_auxiliar == "02923" ~	"CAINSA CAMPO 1",
      localidad_auxiliar == "02925" ~	"CAINSA",
      localidad_auxiliar == "02927" ~	"PASO CAMPAMENTO",
      localidad_auxiliar == "02928" ~	"CATALÁN GRANDE",
      localidad_auxiliar == "02929" ~	"CATALÁN VOLCÁN",
      localidad_auxiliar == "02930" ~	"CHIFLERO",
      localidad_auxiliar == "02931" ~	"COLONIA ESTRELLA",
      localidad_auxiliar == "02933" ~	"COLONIA VIÑAR",
      localidad_auxiliar == "02934" ~	"DIEGO LAMAS",
      localidad_auxiliar == "02935" ~	"ESTIVA",
      localidad_auxiliar == "02936" ~	"FAGÚNDEZ",
      localidad_auxiliar == "02937" ~	"GRANJA PERRONI",
      localidad_auxiliar == "02938" ~	"LENGUAZO",
      localidad_auxiliar == "02940" ~	"PALMA SOLA",
      localidad_auxiliar == "02941" ~	"PAREDÓN",
      localidad_auxiliar == "02942" ~	"PASO DE LA CRUZ",
      localidad_auxiliar == "02943" ~	"PASO DEL LEÓN",
      localidad_auxiliar == "02944" ~	"PASO DE RAMOS",
      localidad_auxiliar == "02945" ~	"PASO FARIAS",
      localidad_auxiliar == "02946" ~	"PASO POTRERO",
      localidad_auxiliar == "02947" ~	"PATITAS",
      localidad_auxiliar == "02948" ~	"PIEDRA PINTADA",
      localidad_auxiliar == "02949" ~	"PINTADO GRANDE",
      localidad_auxiliar == "02950" ~	"PUNTAS DE TRES CRUCES",
      localidad_auxiliar == "02951" ~	"RICARDINHO",
      localidad_auxiliar == "02952" ~	"RINCON DE PACHECO",
      localidad_auxiliar == "02953" ~	"RINCÓN DE PINTADO",
      localidad_auxiliar == "02954" ~	"SARANDÍ DE CUARÓ",
      localidad_auxiliar == "02955" ~	"SARANDÍ DE YACUY",
      localidad_auxiliar == "02956" ~	"TARUMAN",
      localidad_auxiliar == "02957" ~	"TOPADOR",
      localidad_auxiliar == "02958" ~	"CERRO EJIDO",
      localidad_auxiliar == "02959" ~	"COLONIA PALMA",
      localidad_auxiliar == "02960" ~	"ZANJA ARUERA",
      localidad_auxiliar == "02961" ~	"CERRO SIGNORELLI",
      localidad_auxiliar == "02962" ~	"CERRO SAN EUGENIO",
      localidad_auxiliar == "02963" ~	"CALNU",
      localidad_auxiliar == "02964" ~	"SAN GREGORIO",
      localidad_auxiliar == "02965" ~	"MASOLLER",
      localidad_auxiliar == "03024" ~	"PASO CARRASCO",
      localidad_auxiliar == "03221" ~	"LAS PIEDRAS",
      localidad_auxiliar == "03320" ~	"CANELONES",
      localidad_auxiliar == "03321" ~	"LA PAZ",
      localidad_auxiliar == "03322" ~	"PANDO",
      localidad_auxiliar == "03323" ~	"SANTA LUCIA",
      localidad_auxiliar == "03324" ~	"PIEDRAS DE AFILAR",
      localidad_auxiliar == "03325" ~	"CUMBRES DE CARRASCO",
      localidad_auxiliar == "03326" ~	"HARAS DEL LAGO",
      localidad_auxiliar == "03327" ~	"QUINTA LOS HORNEROS",
      localidad_auxiliar == "03328" ~	"LAS HIGUERITAS",
      localidad_auxiliar == "03329" ~	"SOFIA SANTOS",
      localidad_auxiliar == "03330" ~	"HUERTA DE LOS HORNEROS",
      localidad_auxiliar == "03332" ~	"LA HORQUETA",
      localidad_auxiliar == "03333" ~	"BARRIO LA CINACINA",
      localidad_auxiliar == "03334" ~	"CARRASCO DEL SAUCE",
      localidad_auxiliar == "03335" ~	"CUCHILLA DE SIERRA",
      localidad_auxiliar == "03336" ~	"EL COLORADO",
      localidad_auxiliar == "03337" ~	"LAS BRUJAS",
      localidad_auxiliar == "03338" ~	"CUCHILLA VERDE",
      localidad_auxiliar == "03339" ~	"COSTA DE PEDERNAL",
      localidad_auxiliar == "03340" ~	"CERAMICAS SALINAS",
      localidad_auxiliar == "03341" ~	"BARRA DE PEDRERA",
      localidad_auxiliar == "03342" ~	"VILLA LOS NOLITOS",
      localidad_auxiliar == "03343" ~	"18 DE MAYO",
      localidad_auxiliar == "03344" ~	"LA JUANA",
      localidad_auxiliar == "03345" ~	"CASAHONA",
      localidad_auxiliar == "03346" ~	"LA CANTERA",
      localidad_auxiliar == "03347" ~	"EL MONTE",
      localidad_auxiliar == "03348" ~	"LOS TILOS",
      localidad_auxiliar == "03349" ~	"PILAR DE LOS HORNEROS",
      localidad_auxiliar == "03421" ~	"PROGRESO",
      localidad_auxiliar == "03422" ~	"SAN RAMON",
      localidad_auxiliar == "03521" ~	"BARROS BLANCOS",
      localidad_auxiliar == "03522" ~	"FRACC. CNO. MALDONADO",
      localidad_auxiliar == "03523" ~	"COLONIA NICOLICH",
      localidad_auxiliar == "03524" ~	"JOAQUIN SUAREZ",
      localidad_auxiliar == "03526" ~	"SANTA ROSA",
      localidad_auxiliar == "03527" ~	"SAUCE",
      localidad_auxiliar == "03528" ~	"TALA",
      localidad_auxiliar == "03529" ~	"VILLA CRESPO Y SAN ANDRES",
      localidad_auxiliar == "03530" ~	"FRACC. CNO. ANDALUZ Y R.84",
      localidad_auxiliar == "03621" ~	"ATLANTIDA",
      localidad_auxiliar == "03622" ~	"ESTACION ATLANTIDA",
      localidad_auxiliar == "03624" ~	"LOS CERRILLOS",
      localidad_auxiliar == "03625" ~	"EMPALME OLMOS",
      localidad_auxiliar == "03626" ~	"MIGUES",
      localidad_auxiliar == "03627" ~	"PARQUE DEL PLATA",
      localidad_auxiliar == "03628" ~	"SAN BAUTISTA",
      localidad_auxiliar == "03629" ~	"SAN JACINTO",
      localidad_auxiliar == "03630" ~	"DR. FRANCISCO SOCA",
      localidad_auxiliar == "03631" ~	"TOLEDO",
      localidad_auxiliar == "03632" ~	"MONTES",
      localidad_auxiliar == "03633" ~	"SAN JOSE DE CARRASCO",
      localidad_auxiliar == "03634" ~	"FRACC. SOBRE RUTA 74",
      localidad_auxiliar == "03721" ~	"AGUAS CORRIENTES",
      localidad_auxiliar == "03722" ~	"BARRA DE CARRASCO",
      localidad_auxiliar == "03723" ~	"JUANICO",
      localidad_auxiliar == "03724" ~	"LA FLORESTA",
      localidad_auxiliar == "03725" ~	"ESTACION LA FLORESTA",
      localidad_auxiliar == "03726" ~	"LAS TOSCAS",
      localidad_auxiliar == "03727" ~	"PARQUE CARRASCO",
      localidad_auxiliar == "03729" ~	"SALINAS",
      localidad_auxiliar == "03730" ~	"SAN ANTONIO",
      localidad_auxiliar == "03732" ~	"SOLYMAR",
      localidad_auxiliar == "03733" ~	"VILLA AEROPARQUE",
      localidad_auxiliar == "03822" ~	"CAMINO A LA CADENA",
      localidad_auxiliar == "03823" ~	"CASTELLANOS",
      localidad_auxiliar == "03824" ~	"COLONIA BERRO",
      localidad_auxiliar == "03825" ~	"BARRIO COPOLA",
      localidad_auxiliar == "03826" ~	"COSTA AZUL",
      localidad_auxiliar == "03827" ~	"COSTA Y GUILLAMON",
      localidad_auxiliar == "03828" ~	"EL PINAR",
      localidad_auxiliar == "03829" ~	"ESTACION MIGUES",
      localidad_auxiliar == "03830" ~	"PINAMAR - PINEPARK",
      localidad_auxiliar == "03831" ~	"LAGOMAR",
      localidad_auxiliar == "03832" ~	"OLMOS",
      localidad_auxiliar == "03833" ~	"PARADA CABRERA",
      localidad_auxiliar == "03834" ~	"SAN LUIS",
      localidad_auxiliar == "03835" ~	"SHANGRILA",
      localidad_auxiliar == "03836" ~	"TOTORAL DEL SAUCE",
      localidad_auxiliar == "03837" ~	"VILLA FELICIDAD",
      localidad_auxiliar == "03838" ~	"VILLA PAZ S.A.",
      localidad_auxiliar == "03839" ~	"VILLA SAN JOSE",
      localidad_auxiliar == "03840" ~	"ESTACION TAPIA",
      localidad_auxiliar == "03880" ~	"VILLA SAN FELIPE",
      localidad_auxiliar == "03881" ~	"VILLA HADITA",
      localidad_auxiliar == "03882" ~	"PASO DE PACHE",
      localidad_auxiliar == "03900" ~	"RURAL/PARAJE RURAL",
      localidad_auxiliar == "03910" ~	"CITY GOLF",
      localidad_auxiliar == "03911" ~	"VIEJO MOLINO SAN BERNARDO",
      localidad_auxiliar == "03912" ~	"ESTANQUE DE PANDO",
      localidad_auxiliar == "03913" ~	"JARDINES DE PANDO",
      localidad_auxiliar == "03914" ~	"PASO ESPINOSA",
      localidad_auxiliar == "03921" ~	"ARAMINDA",
      localidad_auxiliar == "03922" ~	"ARGENTINO",
      localidad_auxiliar == "03923" ~	"BARRA DE LA PEDRERA",
      localidad_auxiliar == "03924" ~	"BARRANCAS COLORADAS",
      localidad_auxiliar == "03925" ~	"BELLO HORIZONTE",
      localidad_auxiliar == "03926" ~	"BIARRITZ",
      localidad_auxiliar == "03927" ~	"BOLIVAR",
      localidad_auxiliar == "03928" ~	"CAMPO MILITAR",
      localidad_auxiliar == "03929" ~	"CAPILLA DE CELLA",
      localidad_auxiliar == "03930" ~	"CAÑADA DE CARDOZO",
      localidad_auxiliar == "03931" ~	"CERRILLOS AL SUR",
      localidad_auxiliar == "03932" ~	"COSTA DE PANDO",
      localidad_auxiliar == "03933" ~	"COSTA DE TALA",
      localidad_auxiliar == "03934" ~	"CRUZ DE LOS CAMINOS",
      localidad_auxiliar == "03935" ~	"CUCHILLA ALTA",
      localidad_auxiliar == "03937" ~	"CUEVA DEL TIGRE",
      localidad_auxiliar == "03938" ~	"ECHEVARRÍA",
      localidad_auxiliar == "03939" ~	"EL BOSQUE",
      localidad_auxiliar == "03940" ~	"EMPALME SAUCE",
      localidad_auxiliar == "03941" ~	"ESTACIÓN MARGAT",
      localidad_auxiliar == "03942" ~	"ESTACION PEDRERA",
      localidad_auxiliar == "03943" ~	"FORTIN DE SANTA ROSA",
      localidad_auxiliar == "03944" ~	"FRACC. PROGRESO",
      localidad_auxiliar == "03945" ~	"INSTITUTO ADVENTISTA",
      localidad_auxiliar == "03946" ~	"JAUREGUIBERRY",
      localidad_auxiliar == "03947" ~	"LA CAPILLA",
      localidad_auxiliar == "03948" ~	"LA LUCHA",
      localidad_auxiliar == "03949" ~	"LA MONTAÑESA",
      localidad_auxiliar == "03950" ~	"LA PALMITA",
      localidad_auxiliar == "03951" ~	"LA PALOMA",
      localidad_auxiliar == "03952" ~	"LA QUERENCIA",
      localidad_auxiliar == "03953" ~	"LOMAS DE SOLYMAR",
      localidad_auxiliar == "03954" ~	"LAS BARRERAS",
      localidad_auxiliar == "03956" ~	"LOS TITANES",
      localidad_auxiliar == "03957" ~	"MARINDIA",
      localidad_auxiliar == "03959" ~	"NEPTUNIA",
      localidad_auxiliar == "03961" ~	"PARADOR TAJES",
      localidad_auxiliar == "03962" ~	"PASO DE LA CADENA",
      localidad_auxiliar == "03963" ~	"PASO DE LA PALOMA",
      localidad_auxiliar == "03964" ~	"PASO DE LAS TOSCAS",
      localidad_auxiliar == "03965" ~	"PASO DEL BOTE",
      localidad_auxiliar == "03966" ~	"PASO PALOMEQUE",
      localidad_auxiliar == "03967" ~	"PASO VILLAR",
      localidad_auxiliar == "03968" ~	"PIEDRA DEL TORO",
      localidad_auxiliar == "03969" ~	"ESTACION PIEDRAS DE AFILAR",
      localidad_auxiliar == "03970" ~	"EL GALEON",
      localidad_auxiliar == "03972" ~	"PUNTAS DE PANTANOSO",
      localidad_auxiliar == "03973" ~	"SAN PEDRO",
      localidad_auxiliar == "03974" ~	"SANTA ANA",
      localidad_auxiliar == "03975" ~	"SANTA LUCIA DEL ESTE",
      localidad_auxiliar == "03976" ~	"SANTOS LUGARES",
      localidad_auxiliar == "03977" ~	"SAUCE DE SOLÍS",
      localidad_auxiliar == "03978" ~	"SEIS HERMANOS",
      localidad_auxiliar == "03979" ~	"VILLA AREJO",
      localidad_auxiliar == "03980" ~	"VILLA ARGENTINA",
      localidad_auxiliar == "03981" ~	"VILLA ENCANTADA",
      localidad_auxiliar == "03982" ~	"VILLA GABI",
      localidad_auxiliar == "03983" ~	"VILLA NUEVA",
      localidad_auxiliar == "03984" ~	"VILLA PORVENIR",
      localidad_auxiliar == "03985" ~	"LA TUNA",
      localidad_auxiliar == "03986" ~	"GUAZU - VIRA",
      localidad_auxiliar == "03987" ~	"COLINAS DE SOLYMAR",
      localidad_auxiliar == "03988" ~	"BARRIO REMANSO",
      localidad_auxiliar == "03989" ~	"VILLA EL TATO",
      localidad_auxiliar == "03990" ~	"VILLA SAN CONO",
      localidad_auxiliar == "03991" ~	"VILLA JUANA",
      localidad_auxiliar == "03992" ~	"COLINAS DE CARRASCO",
      localidad_auxiliar == "03993" ~	"LOMAS DE CARRASCO",
      localidad_auxiliar == "03994" ~	"CARMEL",
      localidad_auxiliar == "03995" ~	"LA ASUNCION",
      localidad_auxiliar == "03996" ~	"QUINTAS DEL BOSQUE",
      localidad_auxiliar == "03997" ~	"LA TAHONA",
      localidad_auxiliar == "03998" ~	"ASENTAMIENTO R.6 KM 24.500",
      localidad_auxiliar == "03999" ~	"MIRADOR DE LA TAHONA",
      localidad_auxiliar == "04020" ~	"MELO",
      localidad_auxiliar == "04521" ~	"FRAILE MUERTO",
      localidad_auxiliar == "04522" ~	"RIO BRANCO",
      localidad_auxiliar == "04621" ~	"TUPAMBAE",
      localidad_auxiliar == "04721" ~	"ISIDORO NOBLIA",
      localidad_auxiliar == "04821" ~	"ACEGUA",
      localidad_auxiliar == "04822" ~	"BAÑADO DE MEDINA",
      localidad_auxiliar == "04823" ~	"CENTURION",
      localidad_auxiliar == "04824" ~	"CERRO DE LAS CUENTAS",
      localidad_auxiliar == "04825" ~	"HIPODROMO",
      localidad_auxiliar == "04826" ~	"PASO PEREIRA",
      localidad_auxiliar == "04827" ~	"PLACIDO ROSAS",
      localidad_auxiliar == "04828" ~	"TOLEDO",
      localidad_auxiliar == "04829" ~	"TRES ISLAS",
      localidad_auxiliar == "04830" ~	"POBLADO URUGUAY",
      localidad_auxiliar == "04900" ~	"RURAL/PARAJE RURAL",
      localidad_auxiliar == "04921" ~	"AGUIRRE",
      localidad_auxiliar == "04922" ~	"AMARILLO",
      localidad_auxiliar == "04923" ~	"ARBOLITO",
      localidad_auxiliar == "04924" ~	"AREVALO",
      localidad_auxiliar == "04925" ~	"BAÑADO DE SATURNINO",
      localidad_auxiliar == "04926" ~	"BUENA VISTA",
      localidad_auxiliar == "04927" ~	"CALERA DE RECALDE",
      localidad_auxiliar == "04928" ~	"CAMPAMENTO",
      localidad_auxiliar == "04929" ~	"CAÑADA GRANDE",
      localidad_auxiliar == "04930" ~	"CASERIO LAS CAÑAS",
      localidad_auxiliar == "04931" ~	"CAÑITAS",
      localidad_auxiliar == "04932" ~	"COIMBRA",
      localidad_auxiliar == "04933" ~	"CRUZ DE PIEDRA",
      localidad_auxiliar == "04934" ~	"CUCHILLA CAMBOTA",
      localidad_auxiliar == "04935" ~	"CUCHILLA DE MELO",
      localidad_auxiliar == "04936" ~	"CUCHILLA DEL CARMEN",
      localidad_auxiliar == "04937" ~	"CUCHILLA PERALTA",
      localidad_auxiliar == "04938" ~	"DURAZNERO",
      localidad_auxiliar == "04939" ~	"ESCUELA DE AGRONOMÍA",
      localidad_auxiliar == "04940" ~	"ESPERANZA",
      localidad_auxiliar == "04941" ~	"GANEN",
      localidad_auxiliar == "04942" ~	"GARAO",
      localidad_auxiliar == "04943" ~	"GETULIO VARGAS",
      localidad_auxiliar == "04944" ~	"GUAZUNAMBÍ",
      localidad_auxiliar == "04945" ~	"LA CORONILLA",
      localidad_auxiliar == "04946" ~	"LA GLORIA",
      localidad_auxiliar == "04948" ~	"LA PEDRERA",
      localidad_auxiliar == "04949" ~	"LAGO MERIN",
      localidad_auxiliar == "04950" ~	"LAGUNA DEL JUNCO",
      localidad_auxiliar == "04951" ~	"LAS LIMAS",
      localidad_auxiliar == "04952" ~	"MANGRULLO",
      localidad_auxiliar == "04953" ~	"MARÍA ISABEL",
      localidad_auxiliar == "04954" ~	"MEDEROS",
      localidad_auxiliar == "04955" ~	"MONTECITO",
      localidad_auxiliar == "04956" ~	"NANDO",
      localidad_auxiliar == "04957" ~	"PASO DE ALMADA",
      localidad_auxiliar == "04958" ~	"PASO DE LAS TROPAS",
      localidad_auxiliar == "04959" ~	"PASO DEL CENTURIÓN",
      localidad_auxiliar == "04960" ~	"PASO DE MELO",
      localidad_auxiliar == "04961" ~	"PICADA DE SALOMÉ",
      localidad_auxiliar == "04962" ~	"PIEDRA ALTA",
      localidad_auxiliar == "04963" ~	"PIÑEIRO",
      localidad_auxiliar == "04964" ~	"PUENTE DEL CHUY",
      localidad_auxiliar == "04965" ~	"PUNTAS DE TACUARÍ",
      localidad_auxiliar == "04966" ~	"QUEBRACHO",
      localidad_auxiliar == "04967" ~	"RAAB ARROCERA",
      localidad_auxiliar == "04968" ~	"RAMON TRIGO",
      localidad_auxiliar == "04969" ~	"RINCÓN DE PAIVA",
      localidad_auxiliar == "04970" ~	"RINCÓN DE CONTRERAS",
      localidad_auxiliar == "04971" ~	"RINCÓN DE LOS CORONELES",
      localidad_auxiliar == "04972" ~	"RINCÓN DE LOS MONTANA",
      localidad_auxiliar == "04973" ~	"RINCÓN DE LOS OLIVERA",
      localidad_auxiliar == "04974" ~	"RINCÓN DE PY",
      localidad_auxiliar == "04975" ~	"RODRÍGUEZ",
      localidad_auxiliar == "04976" ~	"SAN DIEGO",
      localidad_auxiliar == "04977" ~	"SAN SERVANDO",
      localidad_auxiliar == "04978" ~	"SANCHEZ",
      localidad_auxiliar == "04979" ~	"SARANDÍ DE ACEGUÁ",
      localidad_auxiliar == "04980" ~	"SOSA",
      localidad_auxiliar == "04981" ~	"SOTO GORO",
      localidad_auxiliar == "04982" ~	"TRES BOLICHES",
      localidad_auxiliar == "04983" ~	"BARRIO LOPEZ BENITEZ",
      localidad_auxiliar == "04984" ~	"PUNTAS DE MINAS",
      localidad_auxiliar == "04985" ~	"CASARONE",
      localidad_auxiliar == "04986" ~	"PAJARO AZUL",
      localidad_auxiliar == "04987" ~	"BARRIO LA VINCHUCA",
      localidad_auxiliar == "04988" ~	"ARACHANIA",
      localidad_auxiliar == "04989" ~	"ÑANGAPIRE",
      localidad_auxiliar == "04990" ~	"LA MICAELA",
      localidad_auxiliar == "05320" ~	"COLONIA DEL SACRAMENTO",
      localidad_auxiliar == "05321" ~	"CARMELO",
      localidad_auxiliar == "05323" ~	"JUAN LACAZE",
      localidad_auxiliar == "05421" ~	"NUEVA HELVECIA",
      localidad_auxiliar == "05422" ~	"ROSARIO",
      localidad_auxiliar == "05423" ~	"NUEVA PALMIRA",
      localidad_auxiliar == "05452" ~	"PASO ANTOLIN",
      localidad_auxiliar == "05521" ~	"OMBUES DE LAVALLE",
      localidad_auxiliar == "05522" ~	"TARARIRAS",
      localidad_auxiliar == "05621" ~	"COLONIA VALDENSE",
      localidad_auxiliar == "05622" ~	"FLORENCIO SANCHEZ",
      localidad_auxiliar == "05721" ~	"CONCHILLAS",
      localidad_auxiliar == "05722" ~	"CASERIO EL CERRO",
      localidad_auxiliar == "05723" ~	"EL GENERAL",
      localidad_auxiliar == "05724" ~	"LA PAZ",
      localidad_auxiliar == "05726" ~	"RIACHUELO",
      localidad_auxiliar == "05727" ~	"JUAN CARLOS CASEROS",
      localidad_auxiliar == "05791" ~	"ISMAEL CORTINAS",
      localidad_auxiliar == "05798" ~	"AGRACIADA",
      localidad_auxiliar == "05821" ~	"BOCA DEL ROSARIO",
      localidad_auxiliar == "05822" ~	"BUENA HORA",
      localidad_auxiliar == "05823" ~	"COLONIA ESTRELLA",
      localidad_auxiliar == "05824" ~	"CUFRE",
      localidad_auxiliar == "05825" ~	"EL CAÑO",
      localidad_auxiliar == "05826" ~	"EL SEMILLERO",
      localidad_auxiliar == "05827" ~	"ESTANZUELA",
      localidad_auxiliar == "05828" ~	"JUAN GONZÁLEZ R 21KM 243 A 246",
      localidad_auxiliar == "05829" ~	"CERROS DE SAN JUAN",
      localidad_auxiliar == "05830" ~	"MARTÍN CHICO",
      localidad_auxiliar == "05831" ~	"MIGUELETE",
      localidad_auxiliar == "05832" ~	"PASTOREO",
      localidad_auxiliar == "05833" ~	"PIEDRA DE LOS INDIOS",
      localidad_auxiliar == "05834" ~	"SAN JUAN",
      localidad_auxiliar == "05835" ~	"CARMELO GOLF",
      localidad_auxiliar == "05836" ~	"MONTES DEL PLATA",
      localidad_auxiliar == "05837" ~	"LA CONCORDIA",
      localidad_auxiliar == "05838" ~	"CERRO DE LAS ARMAS",
      localidad_auxiliar == "05839" ~	"CAMPO CHICO",
      localidad_auxiliar == "05900" ~	"RURAL/PARAJE RURAL",
      localidad_auxiliar == "05923" ~	"CAMPANA",
      localidad_auxiliar == "05924" ~	"ARTILLEROS",
      localidad_auxiliar == "05925" ~	"BARKER",
      localidad_auxiliar == "05926" ~	"BARRANCAS COLORADAS",
      localidad_auxiliar == "05927" ~	"EL ENSUEÑO",
      localidad_auxiliar == "05928" ~	"BARRIO HIPÓDROMO",
      localidad_auxiliar == "05929" ~	"BELGRANO NORTE",
      localidad_auxiliar == "05930" ~	"BELGRANO SUR",
      localidad_auxiliar == "05931" ~	"BLANCA ARENA",
      localidad_auxiliar == "05932" ~	"BRISAS DEL PLATA",
      localidad_auxiliar == "05934" ~	"CANTERAS DEL RIACHUELO",
      localidad_auxiliar == "05936" ~	"CERROS NEGROS",
      localidad_auxiliar == "05937" ~	"COLONIA ARRUE",
      localidad_auxiliar == "05938" ~	"COLONIA COSMOPOLITA",
      localidad_auxiliar == "05939" ~	"COLONIA SARANDÍ",
      localidad_auxiliar == "05940" ~	"PARAJE MINUANO",
      localidad_auxiliar == "05941" ~	"COSTA DE COLLA AL ESTE",
      localidad_auxiliar == "05942" ~	"COSTA DE COLLA AL NORTE",
      localidad_auxiliar == "05943" ~	"COSTA DE NAVARRO",
      localidad_auxiliar == "05944" ~	"EL BAÑADO",
      localidad_auxiliar == "05945" ~	"EL CUADRO",
      localidad_auxiliar == "05946" ~	"JUAN GONZÁLEZ",
      localidad_auxiliar == "05947" ~	"LA LAGUNA",
      localidad_auxiliar == "05948" ~	"LAGUNITA",
      localidad_auxiliar == "05949" ~	"LOMAS DE CARMELO",
      localidad_auxiliar == "05950" ~	"LAS FLORES",
      localidad_auxiliar == "05951" ~	"LOS PINOS",
      localidad_auxiliar == "05952" ~	"MIGUELETE DE CONCHILLAS",
      localidad_auxiliar == "05953" ~	"MINAS DE TALCO DE NARANCIO",
      localidad_auxiliar == "05954" ~	"MOLLES DE MIGUELETE",
      localidad_auxiliar == "05955" ~	"CHICO TORINO",
      localidad_auxiliar == "05956" ~	"LA HORQUETA",
      localidad_auxiliar == "05957" ~	"PASO HOSPITAL",
      localidad_auxiliar == "05958" ~	"PASO QUICHO",
      localidad_auxiliar == "05959" ~	"PLAYA AZUL",
      localidad_auxiliar == "05960" ~	"PLAYA BRITOPOLIS",
      localidad_auxiliar == "05961" ~	"PLAYA PARANT",
      localidad_auxiliar == "05962" ~	"PLAYA FOMENTO",
      localidad_auxiliar == "05963" ~	"POLANCO",
      localidad_auxiliar == "05964" ~	"PUERTO INGLES",
      localidad_auxiliar == "05965" ~	"PUERTO ROSARIO",
      localidad_auxiliar == "05966" ~	"PUNTA DE ARENALES",
      localidad_auxiliar == "05967" ~	"PUNTAS DE JUAN GONZÁLEZ",
      localidad_auxiliar == "05968" ~	"PUNTAS DEL ROSARIO (ZUNIN)",
      localidad_auxiliar == "05969" ~	"RADIAL HERNANDEZ",
      localidad_auxiliar == "05970" ~	"RADIAL ROSARIO",
      localidad_auxiliar == "05971" ~	"RESGUARDO CUFRÉ",
      localidad_auxiliar == "05972" ~	"ROSARIO Y COLLA",
      localidad_auxiliar == "05973" ~	"RUTA 21 KM 202",
      localidad_auxiliar == "05974" ~	"SAN LUIS",
      localidad_auxiliar == "05975" ~	"SAN LUIS SÁNCHEZ",
      localidad_auxiliar == "05977" ~	"SAN PEDRO",
      localidad_auxiliar == "05978" ~	"SAN ROQUE",
      localidad_auxiliar == "05979" ~	"SANTA ANA",
      localidad_auxiliar == "05980" ~	"SANTA REGINA",
      localidad_auxiliar == "05981" ~	"SANTA ROSA",
      localidad_auxiliar == "05982" ~	"SARANDÍ CAMPANA",
      localidad_auxiliar == "05983" ~	"TERMINAL - ARTILLEROS",
      localidad_auxiliar == "05984" ~	"TRES ESQUINAS",
      localidad_auxiliar == "05985" ~	"VÍBORAS",
      localidad_auxiliar == "05986" ~	"VÍBORAS OESTE",
      localidad_auxiliar == "05987" ~	"ZAGARZAZU",
      localidad_auxiliar == "05989" ~	"ARRIVILLAGA",
      localidad_auxiliar == "05992" ~	"PALO SOLO",
      localidad_auxiliar == "05993" ~	"EL FARO",
      localidad_auxiliar == "05994" ~	"LAGUNA DE LOS PATOS",
      localidad_auxiliar == "05996" ~	"JUAN JACKSON",
      localidad_auxiliar == "05997" ~	"PUEBLO GIL",
      localidad_auxiliar == "05998" ~	"CERRO CARMELO",
      localidad_auxiliar == "05999" ~	"EL QUINTON",
      localidad_auxiliar == "06020" ~	"DURAZNO",
      localidad_auxiliar == "06421" ~	"SARANDI DEL YI",
      localidad_auxiliar == "06521" ~	"CARMEN",
      localidad_auxiliar == "06621" ~	"BLANQUILLO",
      localidad_auxiliar == "06622" ~	"LA PALOMA",
      localidad_auxiliar == "06721" ~	"CARLOS REYLES",
      localidad_auxiliar == "06722" ~	"CENTENARIO",
      localidad_auxiliar == "06724" ~	"SANTA BERNARDINA",
      localidad_auxiliar == "06790" ~	"CERRO CHATO",
      localidad_auxiliar == "06821" ~	"BAYGORRIA",
      localidad_auxiliar == "06900" ~	"RURAL/PARAJE RURAL",
      localidad_auxiliar == "06921" ~	"ABELLA",
      localidad_auxiliar == "06922" ~	"AGUAS BUENAS",
      localidad_auxiliar == "06923" ~	"CHILENO",
      localidad_auxiliar == "06924" ~	"PUEBLO DE ALVAREZ",
      localidad_auxiliar == "06925" ~	"BARRANCAS COLORADAS",
      localidad_auxiliar == "06926" ~	"BATOVÍ",
      localidad_auxiliar == "06927" ~	"BELLACO",
      localidad_auxiliar == "06928" ~	"BLANQUILLO AL OESTE",
      localidad_auxiliar == "06929" ~	"CAPILLA FARRUCO",
      localidad_auxiliar == "06930" ~	"CARPINTERÍA",
      localidad_auxiliar == "06931" ~	"CEIBAL",
      localidad_auxiliar == "06932" ~	"CERREZUELO",
      localidad_auxiliar == "06933" ~	"CERRO CONVENTO",
      localidad_auxiliar == "06934" ~	"COSTA DE CUADRA",
      localidad_auxiliar == "06935" ~	"CUCHILLA DE RAMÍREZ",
      localidad_auxiliar == "06936" ~	"DE DIOS",
      localidad_auxiliar == "06937" ~	"EL PESCADO",
      localidad_auxiliar == "06938" ~	"ESTACIÓN CHILENO",
      localidad_auxiliar == "06939" ~	"ESTACIÓN PARISH",
      localidad_auxiliar == "06940" ~	"FELICIANO",
      localidad_auxiliar == "06941" ~	"FONSECA",
      localidad_auxiliar == "06942" ~	"LA ALEGRÍA",
      localidad_auxiliar == "06943" ~	"LA MAZAMORRA",
      localidad_auxiliar == "06944" ~	"LOS AGREGADOS",
      localidad_auxiliar == "06945" ~	"LOS AGÜERO",
      localidad_auxiliar == "06947" ~	"LOS ROJAS",
      localidad_auxiliar == "06948" ~	"MALBAJAR",
      localidad_auxiliar == "06949" ~	"MARÍA CEJAS",
      localidad_auxiliar == "06950" ~	"MOURIÑO",
      localidad_auxiliar == "06951" ~	"OMBUES DE ORIBE",
      localidad_auxiliar == "06952" ~	"PARADA SUR KM 265",
      localidad_auxiliar == "06953" ~	"PASO DEL MEDIO LAS PALMAS",
      localidad_auxiliar == "06954" ~	"PUGLIA",
      localidad_auxiliar == "06955" ~	"PUNTA DE LAS FLORES",
      localidad_auxiliar == "06956" ~	"PUNTAS DE HERRERA",
      localidad_auxiliar == "06957" ~	"REYNOLDS",
      localidad_auxiliar == "06958" ~	"ROJAS",
      localidad_auxiliar == "06959" ~	"ROLÓN",
      localidad_auxiliar == "06960" ~	"ROSSELL Y RIUS",
      localidad_auxiliar == "06961" ~	"RUTA 5 KM 172",
      localidad_auxiliar == "06962" ~	"SALINAS",
      localidad_auxiliar == "06963" ~	"SALINAS CHICO",
      localidad_auxiliar == "06964" ~	"SAN JORGE",
      localidad_auxiliar == "06965" ~	"SAN JOSÉ DE LAS CAÑAS",
      localidad_auxiliar == "06966" ~	"SANDÚ CHICO",
      localidad_auxiliar == "06967" ~	"SARANDÍ DEL RÍO NEGRO",
      localidad_auxiliar == "06971" ~	"LAS CAÑAS",
      localidad_auxiliar == "06972" ~	"LAS PALMAS",
      localidad_auxiliar == "06973" ~	"SARANDÍ DE CUADRA",
      localidad_auxiliar == "07320" ~	"TRINIDAD",
      localidad_auxiliar == "07721" ~	"ISMAEL CORTINAS",
      localidad_auxiliar == "07900" ~	"RURAL/PARAJE RURAL",
      localidad_auxiliar == "07921" ~	"ANDRESITO",
      localidad_auxiliar == "07922" ~	"ARENAL CHICO",
      localidad_auxiliar == "07923" ~	"COLONIA ALEMANA",
      localidad_auxiliar == "07924" ~	"EL TOTORAL",
      localidad_auxiliar == "07925" ~	"JUAN JOSE CASTRO",
      localidad_auxiliar == "07926" ~	"PUEBLITO PIEDRA",
      localidad_auxiliar == "07927" ~	"PUEBLO PINTOS",
      localidad_auxiliar == "07928" ~	"PUNTAS DE CHAMANGA",
      localidad_auxiliar == "07929" ~	"PUNTAS DE CORRAL DE PIEDRA",
      localidad_auxiliar == "07930" ~	"PUNTAS DE MARINCHO",
      localidad_auxiliar == "07931" ~	"PUNTAS DE SARANDÍ",
      localidad_auxiliar == "07932" ~	"PUNTAS DEL SAUCE",
      localidad_auxiliar == "07933" ~	"SAN GREGORIO",
      localidad_auxiliar == "07934" ~	"SANTA ADELAIDA",
      localidad_auxiliar == "07935" ~	"SANTA ELENA",
      localidad_auxiliar == "07936" ~	"TALAS DE MACIEL",
      localidad_auxiliar == "07937" ~	"LA CASILLA",
      localidad_auxiliar == "07938" ~	"CERRO COLORADO",
      localidad_auxiliar == "08220" ~	"FLORIDA",
      localidad_auxiliar == "08421" ~	"SARANDI GRANDE",
      localidad_auxiliar == "08521" ~	"CASUPA",
      localidad_auxiliar == "08621" ~	"CARDAL",
      localidad_auxiliar == "08622" ~	"FRAY MARCOS",
      localidad_auxiliar == "08623" ~	"25 DE AGOSTO",
      localidad_auxiliar == "08624" ~	"25 DE MAYO",
      localidad_auxiliar == "08721" ~	"ALEJANDRO GALLINAL",
      localidad_auxiliar == "08722" ~	"CAPILLA DEL SAUCE",
      localidad_auxiliar == "08723" ~	"LA CRUZ",
      localidad_auxiliar == "08724" ~	"NICO PEREZ",
      localidad_auxiliar == "08790" ~	"CERRO CHATO",
      localidad_auxiliar == "08821" ~	"CHAMIZO",
      localidad_auxiliar == "08822" ~	"GOÑI",
      localidad_auxiliar == "08823" ~	"MENDOZA",
      localidad_auxiliar == "08824" ~	"MENDOZA CHICO",
      localidad_auxiliar == "08825" ~	"REBOLEDO",
      localidad_auxiliar == "08896" ~	"VALENTINES",
      localidad_auxiliar == "08900" ~	"RURAL/PARAJE RURAL",
      localidad_auxiliar == "08922" ~	"ARRAYÁN",
      localidad_auxiliar == "08923" ~	"BERRONDO",
      localidad_auxiliar == "08924" ~	"CHAMIZO CHICO",
      localidad_auxiliar == "08925" ~	"CHINGOLAS",
      localidad_auxiliar == "08926" ~	"COLONIA SÁNCHEZ",
      localidad_auxiliar == "08927" ~	"COSTA DE CHAMIZO GRANDE",
      localidad_auxiliar == "08928" ~	"ESTACIÓN PALERMO",
      localidad_auxiliar == "08929" ~	"ESTACIÓN URIOSTE",
      localidad_auxiliar == "08930" ~	"PUEBLO FERRER",
      localidad_auxiliar == "08931" ~	"FRIGORÍFICO MODELO",
      localidad_auxiliar == "08933" ~	"INDEPENDENCIA",
      localidad_auxiliar == "08934" ~	"JUNCAL",
      localidad_auxiliar == "08935" ~	"LAS CHILCAS",
      localidad_auxiliar == "08936" ~	"MANSAVILLAGRA",
      localidad_auxiliar == "08937" ~	"MOLLES DEL PESCADO",
      localidad_auxiliar == "08938" ~	"MONTECORAL",
      localidad_auxiliar == "08939" ~	"PASO DE LOS NOVILLOS",
      localidad_auxiliar == "08940" ~	"PINTADO",
      localidad_auxiliar == "08941" ~	"POLANCO DEL YI",
      localidad_auxiliar == "08942" ~	"PUEBLITO DE LAS ROSAS",
      localidad_auxiliar == "08943" ~	"PUEBLO DE LOS MOROCHOS",
      localidad_auxiliar == "08944" ~	"PUNTAS DE MACIEL",
      localidad_auxiliar == "08945" ~	"PUNTAS DE MANSAVILLAGRA",
      localidad_auxiliar == "08946" ~	"PUNTAS DE SARANDÍ",
      localidad_auxiliar == "08947" ~	"SAN PEDRO DEL TIMOTE",
      localidad_auxiliar == "08948" ~	"TALITA",
      localidad_auxiliar == "08950" ~	"VILLA HÍPICA",
      localidad_auxiliar == "08951" ~	"VILLA VIEJA",
      localidad_auxiliar == "08991" ~	"ILLESCAS",
      localidad_auxiliar == "08992" ~	"CASERIO LA FUNDACION",
      localidad_auxiliar == "08993" ~	"LA MACANA",
      localidad_auxiliar == "08994" ~	"ESTACION CAPILLA DEL SAUCE",
      localidad_auxiliar == "08995" ~	"SAN GABRIEL",
      localidad_auxiliar == "09220" ~	"MINAS",
      localidad_auxiliar == "09521" ~	"JOSE BATLLE Y ORDO?ÂÑEZ",
      localidad_auxiliar == "09522" ~	"JOSE PEDRO VARELA",
      localidad_auxiliar == "09621" ~	"MARISCALA",
      localidad_auxiliar == "09622" ~	"SOLIS DE MATAOJO",
      localidad_auxiliar == "09721" ~	"PIRARAJA",
      localidad_auxiliar == "09722" ~	"ZAPICAN",
      localidad_auxiliar == "09821" ~	"COLON",
      localidad_auxiliar == "09822" ~	"COSTAS DEL SOLDADO",
      localidad_auxiliar == "09823" ~	"ARAMENDIA",
      localidad_auxiliar == "09824" ~	"RUTA 40 KM 27,5",
      localidad_auxiliar == "09900" ~	"RURAL/PARAJE RURAL",
      localidad_auxiliar == "09921" ~	"AGUAS BLANCAS",
      localidad_auxiliar == "09922" ~	"ALONSO",
      localidad_auxiliar == "09923" ~	"ANDREONI",
      localidad_auxiliar == "09924" ~	"BLANES VIALE",
      localidad_auxiliar == "09925" ~	"CARNALES",
      localidad_auxiliar == "09926" ~	"CERRO PELADO",
      localidad_auxiliar == "09927" ~	"COSTAS DEL LENGUAZO",
      localidad_auxiliar == "09928" ~	"19 DE JUNIO",
      localidad_auxiliar == "09929" ~	"EL SOLDADO",
      localidad_auxiliar == "09930" ~	"ESTACIÓN ORTIZ",
      localidad_auxiliar == "09931" ~	"ESTACION SOLIS",
      localidad_auxiliar == "09932" ~	"GAETAN",
      localidad_auxiliar == "09933" ~	"GODOY",
      localidad_auxiliar == "09934" ~	"HIGUERITAS",
      localidad_auxiliar == "09935" ~	"LA PLATA",
      localidad_auxiliar == "09936" ~	"LADRILLOS",
      localidad_auxiliar == "09937" ~	"LAS ACHIRAS",
      localidad_auxiliar == "09938" ~	"MARCO DE LOS REYES",
      localidad_auxiliar == "09939" ~	"MARMARAJÁ",
      localidad_auxiliar == "09940" ~	"MOLLES DE GUTIÉRREZ",
      localidad_auxiliar == "09941" ~	"POBLADO LARROSA",
      localidad_auxiliar == "09942" ~	"POLANCO NORTE",
      localidad_auxiliar == "09943" ~	"POLANCO SUR",
      localidad_auxiliar == "09944" ~	"PUNTAS DE BARRIGA NEGRA",
      localidad_auxiliar == "09945" ~	"PUNTAS DE SANTA LUCÍA",
      localidad_auxiliar == "09946" ~	"RETAMOSA",
      localidad_auxiliar == "09947" ~	"RINCÓN DE MARISCALA",
      localidad_auxiliar == "09948" ~	"RINCÓN DE CEBOLLATÍ",
      localidad_auxiliar == "09949" ~	"RUTA 40 KM 25",
      localidad_auxiliar == "09950" ~	"SALUS",
      localidad_auxiliar == "09951" ~	"SARANDÍ DE GUTIÉRREZ",
      localidad_auxiliar == "09952" ~	"TAPES CHICO",
      localidad_auxiliar == "09953" ~	"TAPES GRANDE",
      localidad_auxiliar == "09954" ~	"VELÁZQUEZ",
      localidad_auxiliar == "09955" ~	"VILLA DEL ROSARIO",
      localidad_auxiliar == "09956" ~	"VILLA SERRANA",
      localidad_auxiliar == "09957" ~	"BARRIO LA CORONILLA - ANCAP",
      localidad_auxiliar == "09958" ~	"SAN FRANCISCO DE LAS SIERRAS",
      localidad_auxiliar == "09991" ~	"ILLESCAS",
      localidad_auxiliar == "10020" ~	"MALDONADO",
      localidad_auxiliar == "10021" ~	"PUNTA DEL ESTE",
      localidad_auxiliar == "10022" ~	"PUNTA BALLENA",
      localidad_auxiliar == "10321" ~	"SAN CARLOS",
      localidad_auxiliar == "10521" ~	"AIGUA",
      localidad_auxiliar == "10522" ~	"PAN DE AZUCAR",
      localidad_auxiliar == "10523" ~	"PIRIAPOLIS",
      localidad_auxiliar == "10821" ~	"CERRO PELADO",
      localidad_auxiliar == "10822" ~	"GARZON",
      localidad_auxiliar == "10823" ~	"GERONA  - KM 110",
      localidad_auxiliar == "10824" ~	"LA SIERRA",
      localidad_auxiliar == "10825" ~	"LAS FLORES - ESTACION",
      localidad_auxiliar == "10826" ~	"LOS TALAS",
      localidad_auxiliar == "10827" ~	"NUEVA CARRARA",
      localidad_auxiliar == "10828" ~	"SOLIS",
      localidad_auxiliar == "10829" ~	"PUEBLO SOLIS",
      localidad_auxiliar == "10830" ~	"PINARES - LAS DELICIAS",
      localidad_auxiliar == "10831" ~	"CHIHUAHUA",
      localidad_auxiliar == "10832" ~	"VILLA DELIA",
      localidad_auxiliar == "10833" ~	"SAN RAFAEL - EL PLACER",
      localidad_auxiliar == "10900" ~	"RURAL/PARAJE RURAL",
      localidad_auxiliar == "10919" ~	"ISLAS",
      localidad_auxiliar == "10921" ~	"ABRA DE CASTELLANOS",
      localidad_auxiliar == "10922" ~	"ABRA DE PERDOMO",
      localidad_auxiliar == "10923" ~	"BARRA DEL SAUCE",
      localidad_auxiliar == "10924" ~	"BARRIO HIPODROMO",
      localidad_auxiliar == "10925" ~	"LOS AROMOS",
      localidad_auxiliar == "10926" ~	"BELLA VISTA",
      localidad_auxiliar == "10927" ~	"BUENOS AIRES",
      localidad_auxiliar == "10928" ~	"CALERA RAMOS",
      localidad_auxiliar == "10929" ~	"CANTERAS DE MARELLI",
      localidad_auxiliar == "10930" ~	"CARAPÉ",
      localidad_auxiliar == "10931" ~	"CERROS AZULES",
      localidad_auxiliar == "10932" ~	"CORONILLA",
      localidad_auxiliar == "10934" ~	"EL CHORRO",
      localidad_auxiliar == "10935" ~	"EL EDEN",
      localidad_auxiliar == "10936" ~	"EL TESORO",
      localidad_auxiliar == "10937" ~	"ESTACIÓN JOSÉ IGNACIO",
      localidad_auxiliar == "10939" ~	"FARO JOSE IGNACIO",
      localidad_auxiliar == "10940" ~	"GREGORIO AZNAREZ",
      localidad_auxiliar == "10941" ~	"GUARDIA VIEJA",
      localidad_auxiliar == "10942" ~	"LA BARRA",
      localidad_auxiliar == "10943" ~	"LA FALDA",
      localidad_auxiliar == "10944" ~	"LA CAPUERA",
      localidad_auxiliar == "10945" ~	"LAS FLORES",
      localidad_auxiliar == "10946" ~	"LOS CEIBOS",
      localidad_auxiliar == "10947" ~	"MANANTIALES",
      localidad_auxiliar == "10948" ~	"OCEAN PARK",
      localidad_auxiliar == "10949" ~	"PARTIDO NORTE",
      localidad_auxiliar == "10950" ~	"PARTIDO OESTE",
      localidad_auxiliar == "10951" ~	"PASO DE LOS TALAS",
      localidad_auxiliar == "10952" ~	"PICADA TOLOSA",
      localidad_auxiliar == "10953" ~	"PLAYA GRANDE",
      localidad_auxiliar == "10954" ~	"PLAYA HERMOSA",
      localidad_auxiliar == "10955" ~	"PLAYA VERDE",
      localidad_auxiliar == "10958" ~	"PUNTA COLORADA",
      localidad_auxiliar == "10959" ~	"PUNTA NEGRA",
      localidad_auxiliar == "10960" ~	"PUNTAS DE SAN IGNACIO",
      localidad_auxiliar == "10961" ~	"RINCÓN DE LOS NÚÑEZ",
      localidad_auxiliar == "10962" ~	"RUTA 37 Y 9 - SIERRAS DEL TIROL",
      localidad_auxiliar == "10963" ~	"RUTA 73 KM 101 A 102",
      localidad_auxiliar == "10964" ~	"RUTA 9 KM 86",
      localidad_auxiliar == "10965" ~	"SALAMANCA",
      localidad_auxiliar == "10966" ~	"SAN JUAN DEL ESTE",
      localidad_auxiliar == "10968" ~	"SAUCE DE AIGUÁ",
      localidad_auxiliar == "10969" ~	"SAUCE DE PORTEZUELO",
      localidad_auxiliar == "10972" ~	"PUEBLO SAN VICENTE",
      localidad_auxiliar == "10973" ~	"BALNEARIO BUENOS AIRES",
      localidad_auxiliar == "10974" ~	"DE LOBOS",
      localidad_auxiliar == "10980" ~	"LAS CUMBRES",
      localidad_auxiliar == "10981" ~	"LOS CORCHOS",
      localidad_auxiliar == "10982" ~	"SANTA MONICA",
      localidad_auxiliar == "10983" ~	"EDEN ROCK",
      localidad_auxiliar == "10984" ~	"PARQUE MEDINA",
      localidad_auxiliar == "10985" ~	"ARENAS DE JOSE IGNACIO",
      localidad_auxiliar == "10986" ~	"LA SONRISA",
      localidad_auxiliar == "10987" ~	"EL QUIJOTE",
      localidad_auxiliar == "10988" ~	"LAGUNA BLANCA",
      localidad_auxiliar == "10989" ~	"BARRIO FRIGORIFICO",
      localidad_auxiliar == "10990" ~	"PUEBLO MIO",
      localidad_auxiliar == "10991" ~	"LA MORADA DE PUNTA PIEDRAS",
      localidad_auxiliar == "10992" ~	"LA JUANITA",
      localidad_auxiliar == "10993" ~	"MANGRULLOS DE PUNTA PIEDRAS",
      localidad_auxiliar == "10994" ~	"CHACRAS DEL GOLF",
      localidad_auxiliar == "10995" ~	"FASANO LAS PIEDRAS",
      localidad_auxiliar == "10996" ~	"CHAKRAS DEL ESTE",
      localidad_auxiliar == "10997" ~	"CHACRAS DE MEDELLIN",
      localidad_auxiliar == "10998" ~	"LAGUNA DE LOS CISNES",
      localidad_auxiliar == "11120" ~	"PAYSANDU",
      localidad_auxiliar == "11521" ~	"GUICHON",
      localidad_auxiliar == "11522" ~	"NUEVO PAYSANDU",
      localidad_auxiliar == "11621" ~	"QUEBRACHO",
      localidad_auxiliar == "11695" ~	"TAMBORES",
      localidad_auxiliar == "11721" ~	"LORENZO GEYRES",
      localidad_auxiliar == "11722" ~	"MERINOS",
      localidad_auxiliar == "11723" ~	"PORVENIR",
      localidad_auxiliar == "11797" ~	"ALGORTA",
      localidad_auxiliar == "11821" ~	"ARBOLITO",
      localidad_auxiliar == "11822" ~	"BEISSO",
      localidad_auxiliar == "11823" ~	"CASABLANCA",
      localidad_auxiliar == "11824" ~	"CERRO CHATO",
      localidad_auxiliar == "11825" ~	"CONSTANCIA",
      localidad_auxiliar == "11826" ~	"MORATO",
      localidad_auxiliar == "11827" ~	"PIEDRAS COLORADAS",
      localidad_auxiliar == "11828" ~	"PIÑERA",
      localidad_auxiliar == "11829" ~	"PUNTAS DE BURICAYUPÍ",
      localidad_auxiliar == "11830" ~	"SAN FELIX",
      localidad_auxiliar == "11831" ~	"VILLA MARIA (TIATUCURA)",
      localidad_auxiliar == "11894" ~	"PIEDRA SOLA",
      localidad_auxiliar == "11900" ~	"RURAL/PARAJE RURAL",
      localidad_auxiliar == "11921" ~	"ALONSO",
      localidad_auxiliar == "11922" ~	"ARAUJO",
      localidad_auxiliar == "11924" ~	"BELLA VISTA",
      localidad_auxiliar == "11925" ~	"CAÑADA DEL PUEBLO",
      localidad_auxiliar == "11926" ~	"CHAPICUY",
      localidad_auxiliar == "11927" ~	"DAYMÁN",
      localidad_auxiliar == "11928" ~	"EL CHACO",
      localidad_auxiliar == "11929" ~	"EL EUCALIPTUS",
      localidad_auxiliar == "11930" ~	"EL HORNO",
      localidad_auxiliar == "11931" ~	"ESPERANZA",
      localidad_auxiliar == "11932" ~	"PUEBLO FEDERACION",
      localidad_auxiliar == "11933" ~	"GUAYABOS",
      localidad_auxiliar == "11935" ~	"LA TENTACION",
      localidad_auxiliar == "11936" ~	"LAS FLORES",
      localidad_auxiliar == "11937" ~	"ORGOROSO",
      localidad_auxiliar == "11938" ~	"PALMAR DEL QUEBRACHO",
      localidad_auxiliar == "11939" ~	"PANDULE",
      localidad_auxiliar == "11940" ~	"PASO DE LOS CARROS",
      localidad_auxiliar == "11941" ~	"QUEGUAY CHICO",
      localidad_auxiliar == "11942" ~	"RUTA 90 KM 36",
      localidad_auxiliar == "11943" ~	"SACACHISPAS",
      localidad_auxiliar == "11944" ~	"SAUCE DE ABAJO",
      localidad_auxiliar == "11945" ~	"CUCHILLA DE BURICAYUPI",
      localidad_auxiliar == "11946" ~	"SAUCE DEL QUEGUAY",
      localidad_auxiliar == "11947" ~	"SOTO",
      localidad_auxiliar == "11948" ~	"TOMÁS PAZ",
      localidad_auxiliar == "11949" ~	"VALDEZ",
      localidad_auxiliar == "11950" ~	"ZEBALLOS",
      localidad_auxiliar == "11951" ~	"CHACRAS DE PAYSANDU",
      localidad_auxiliar == "11952" ~	"RIVAS",
      localidad_auxiliar == "11953" ~	"GALLINAL",
      localidad_auxiliar == "11954" ~	"PUNTAS DE ARROYO NEGRO",
      localidad_auxiliar == "11955" ~	"ESTACION PORVENIR",
      localidad_auxiliar == "11956" ~	"CUCHILLA DE FUEGO",
      localidad_auxiliar == "11957" ~	"PUEBLO ALONZO",
      localidad_auxiliar == "11958" ~	"QUEGUAYAR",
      localidad_auxiliar == "11959" ~	"TERMAS DE GUAVIYU",
      localidad_auxiliar == "11960" ~	"TERMAS DE ALMIRON",
      localidad_auxiliar == "11961" ~	"EL ESPINILLO",
      localidad_auxiliar == "12320" ~	"FRAY BENTOS",
      localidad_auxiliar == "12421" ~	"YOUNG",
      localidad_auxiliar == "12621" ~	"NUEVO BERLIN",
      localidad_auxiliar == "12622" ~	"SAN JAVIER",
      localidad_auxiliar == "12722" ~	"BARRIO ANGLO",
      localidad_auxiliar == "12723" ~	"GRECCO",
      localidad_auxiliar == "12792" ~	"MERINOS",
      localidad_auxiliar == "12797" ~	"ALGORTA",
      localidad_auxiliar == "12821" ~	"EL OMBU",
      localidad_auxiliar == "12822" ~	"LOS RANCHOS",
      localidad_auxiliar == "12823" ~	"PASO DE LOS MELLIZOS",
      localidad_auxiliar == "12824" ~	"SARANDI DE NAVARRO",
      localidad_auxiliar == "12825" ~	"SAUCE",
      localidad_auxiliar == "12826" ~	"VILLA GENERAL BORGES",
      localidad_auxiliar == "12827" ~	"VILLA MARIA",
      localidad_auxiliar == "12828" ~	"LAS CAÑAS",
      localidad_auxiliar == "12900" ~	"RURAL/PARAJE RURAL",
      localidad_auxiliar == "12919" ~	"ISLAS",
      localidad_auxiliar == "12921" ~	"ABRIGO",
      localidad_auxiliar == "12923" ~	"BELLACO",
      localidad_auxiliar == "12924" ~	"ISLAS DE ARGUELLO",
      localidad_auxiliar == "12925" ~	"LA ARENA",
      localidad_auxiliar == "12926" ~	"LA FLORIDA",
      localidad_auxiliar == "12927" ~	"LA ILERA",
      localidad_auxiliar == "12928" ~	"LA UNIÓN",
      localidad_auxiliar == "12929" ~	"LIEBIGS",
      localidad_auxiliar == "12930" ~	"LOS ARRAYANES",
      localidad_auxiliar == "12931" ~	"MATAOJO",
      localidad_auxiliar == "12932" ~	"MENAFRA",
      localidad_auxiliar == "12933" ~	"PALMAR GRANDE",
      localidad_auxiliar == "12934" ~	"PASO ARROYO DON ESTEBAN",
      localidad_auxiliar == "12935" ~	"PASO DE LOS COBRES",
      localidad_auxiliar == "12936" ~	"PASO DE SOCA",
      localidad_auxiliar == "12937" ~	"ROLÓN",
      localidad_auxiliar == "12938" ~	"SÁNCHEZ CHICO",
      localidad_auxiliar == "12939" ~	"SÁNCHEZ",
      localidad_auxiliar == "12940" ~	"SANTA ELISA",
      localidad_auxiliar == "12941" ~	"SANTA ROSA",
      localidad_auxiliar == "12942" ~	"SARANDÍ CHICO",
      localidad_auxiliar == "12943" ~	"TRES BOCAS",
      localidad_auxiliar == "12944" ~	"ULESTE",
      localidad_auxiliar == "12945" ~	"TRES QUINTAS",
      localidad_auxiliar == "12946" ~	"BARRIO EL MOLINO",
      localidad_auxiliar == "13020" ~	"RIVERA",
      localidad_auxiliar == "13521" ~	"MINAS DE CORRALES",
      localidad_auxiliar == "13522" ~	"TRANQUERAS",
      localidad_auxiliar == "13621" ~	"VICHADERO",
      localidad_auxiliar == "13721" ~	"SANTA TERESA",
      localidad_auxiliar == "13821" ~	"ABROJAL",
      localidad_auxiliar == "13822" ~	"ARROYO BLANCO",
      localidad_auxiliar == "13823" ~	"PASO ATAQUES",
      localidad_auxiliar == "13824" ~	"CERRO PELADO",
      localidad_auxiliar == "13825" ~	"CERRO CAQUEIRO",
      localidad_auxiliar == "13826" ~	"CHILCA DE CARAGUATÁ",
      localidad_auxiliar == "13827" ~	"CORTUME",
      localidad_auxiliar == "13828" ~	"CURTICEIRAS",
      localidad_auxiliar == "13829" ~	"CUÑAPIRÚ",
      localidad_auxiliar == "13830" ~	"PASO HOSPITAL",
      localidad_auxiliar == "13831" ~	"LAPUENTE",
      localidad_auxiliar == "13832" ~	"LAS FLORES",
      localidad_auxiliar == "13833" ~	"MOIRONES",
      localidad_auxiliar == "13834" ~	"PUNTAS DE CORRALES",
      localidad_auxiliar == "13835" ~	"SAN GREGORIO",
      localidad_auxiliar == "13836" ~	"LA PEDRERA",
      localidad_auxiliar == "13837" ~	"MANDUBÍ",
      localidad_auxiliar == "13838" ~	"LAGUNÓN",
      localidad_auxiliar == "13900" ~	"RURAL/PARAJE RURAL",
      localidad_auxiliar == "13921" ~	"ALBORADA",
      localidad_auxiliar == "13922" ~	"AMARILLO",
      localidad_auxiliar == "13923" ~	"BATOVÍ",
      localidad_auxiliar == "13924" ~	"BERRUTI",
      localidad_auxiliar == "13925" ~	"BLANQUILLOS",
      localidad_auxiliar == "13926" ~	"CAPÓN ALTO",
      localidad_auxiliar == "13927" ~	"CARPINTERÍA",
      localidad_auxiliar == "13928" ~	"CARPINTERÍA DE YAGUARÍ",
      localidad_auxiliar == "13929" ~	"CERRILLADA",
      localidad_auxiliar == "13930" ~	"CERRO ALEGRE",
      localidad_auxiliar == "13931" ~	"CERROS DE LA CALERA",
      localidad_auxiliar == "13932" ~	"CERROS BLANCOS DE CUÑAPIRÚ",
      localidad_auxiliar == "13933" ~	"CORONILLA",
      localidad_auxiliar == "13934" ~	"CORONILLA DE CORRALES",
      localidad_auxiliar == "13935" ~	"CRUZ DE SAN PEDRO",
      localidad_auxiliar == "13936" ~	"CUCHILLA DE TRES CERROS",
      localidad_auxiliar == "13937" ~	"CUCHILLA MANGUERAS",
      localidad_auxiliar == "13938" ~	"GUAVIYÚ",
      localidad_auxiliar == "13939" ~	"LA CHILCA",
      localidad_auxiliar == "13940" ~	"LAGOS DEL NORTE",
      localidad_auxiliar == "13941" ~	"LAURELES",
      localidad_auxiliar == "13942" ~	"MANUEL DÍAZ",
      localidad_auxiliar == "13943" ~	"MASOLLER",
      localidad_auxiliar == "13944" ~	"MINAS DE CUÑAPIRÚ",
      localidad_auxiliar == "13945" ~	"MINAS DE ZAPUCAY",
      localidad_auxiliar == "13946" ~	"PARADA MEDINA",
      localidad_auxiliar == "13947" ~	"PASO DE AMARILLO",
      localidad_auxiliar == "13948" ~	"ESTACION ATAQUES",
      localidad_auxiliar == "13949" ~	"PASO DE SERPA",
      localidad_auxiliar == "13950" ~	"PASO DE TAPADO",
      localidad_auxiliar == "13951" ~	"PASO DEL PARQUE",
      localidad_auxiliar == "13952" ~	"PIEDRAS BLANCAS",
      localidad_auxiliar == "13953" ~	"PLATÓN",
      localidad_auxiliar == "13954" ~	"PUNTAS DE ABROJAL",
      localidad_auxiliar == "13955" ~	"RINCÓN DE RODRÍGUEZ",
      localidad_auxiliar == "13956" ~	"RINCÓN DE ROLAND",
      localidad_auxiliar == "13957" ~	"RINCÓN LOS TRES CERROS",
      localidad_auxiliar == "13958" ~	"RUBIO CHICO",
      localidad_auxiliar == "13959" ~	"SARANDÍ DE RÍO NEGRO",
      localidad_auxiliar == "13960" ~	"SAUZAL",
      localidad_auxiliar == "13961" ~	"VILLA INDART",
      localidad_auxiliar == "13962" ~	"YAGUARÍ",
      localidad_auxiliar == "13963" ~	"ZANJA HONDA 01",
      localidad_auxiliar == "13964" ~	"ZANJA HONDA 02",
      localidad_auxiliar == "13965" ~	"CERRO CHAPEU",
      localidad_auxiliar == "13966" ~	"PUEBLO MADERA/LOS CEIBOS",
      localidad_auxiliar == "14320" ~	"ROCHA",
      localidad_auxiliar == "14421" ~	"CASTILLOS",
      localidad_auxiliar == "14422" ~	"LASCANO",
      localidad_auxiliar == "14521" ~	"CHUY",
      localidad_auxiliar == "14621" ~	"CEBOLLATI",
      localidad_auxiliar == "14622" ~	"VELAZQUEZ",
      localidad_auxiliar == "14721" ~	"18 DE JULIO",
      localidad_auxiliar == "14722" ~	"LA PALOMA",
      localidad_auxiliar == "14723" ~	"SAN LUIS AL MEDIO",
      localidad_auxiliar == "14821" ~	"LA AGUADA Y COSTA AZUL",
      localidad_auxiliar == "14822" ~	"19 DE ABRIL",
      localidad_auxiliar == "14823" ~	"LA CORONILLA",
      localidad_auxiliar == "14825" ~	"PASO BARRANCAS",
      localidad_auxiliar == "14826" ~	"BARRIO PEREIRA",
      localidad_auxiliar == "14900" ~	"RURAL/PARAJE RURAL",
      localidad_auxiliar == "14919" ~	"ISLA",
      localidad_auxiliar == "14921" ~	"AGUAS DULCES",
      localidad_auxiliar == "14922" ~	"ARROZAL VICTORIA",
      localidad_auxiliar == "14923" ~	"BARRA DEL CHUY",
      localidad_auxiliar == "14926" ~	"BARRIO TORRES",
      localidad_auxiliar == "14927" ~	"BUENA VISTA",
      localidad_auxiliar == "14928" ~	"CABO POLONIO",
      localidad_auxiliar == "14929" ~	"CAPACHO",
      localidad_auxiliar == "14930" ~	"CERRO DE LOS ROCHA",
      localidad_auxiliar == "14931" ~	"COSTAS DEL CEIBO",
      localidad_auxiliar == "14932" ~	"CUCHILLA DE GARZÓN",
      localidad_auxiliar == "14933" ~	"CUCHILLA DE INDIA MUERTA",
      localidad_auxiliar == "14934" ~	"EL CANELÓN",
      localidad_auxiliar == "14935" ~	"EL CARACOL",
      localidad_auxiliar == "14936" ~	"EL CEIBO",
      localidad_auxiliar == "14937" ~	"EL CHIMANGO",
      localidad_auxiliar == "14938" ~	"ESTERO PELOTAS ARROCERA",
      localidad_auxiliar == "14939" ~	"ESTIVA DE CHAFALOTE",
      localidad_auxiliar == "14940" ~	"BARRA DE VALIZAS",
      localidad_auxiliar == "14941" ~	"LA ESMERALDA",
      localidad_auxiliar == "14942" ~	"LA PEDRERA",
      localidad_auxiliar == "14943" ~	"LAGUNITA",
      localidad_auxiliar == "14944" ~	"LA GARZAS",
      localidad_auxiliar == "14945" ~	"LOS TACURUSES ARROCERA",
      localidad_auxiliar == "14946" ~	"MENA",
      localidad_auxiliar == "14947" ~	"PALMAR",
      localidad_auxiliar == "14948" ~	"PARALLE",
      localidad_auxiliar == "14949" ~	"PARQUE NAC. DE SANTA TERESA",
      localidad_auxiliar == "14950" ~	"PASO DEL BAÑADO",
      localidad_auxiliar == "14951" ~	"PICADA TECHERA ARROCERA",
      localidad_auxiliar == "14952" ~	"POBLADO CORREA",
      localidad_auxiliar == "14953" ~	"PUERTO DE LOS BOTES",
      localidad_auxiliar == "14954" ~	"QUEBRACHO",
      localidad_auxiliar == "14955" ~	"RINCÓN DE NIETO",
      localidad_auxiliar == "14956" ~	"RINCÓN DE LOS OLIVERA",
      localidad_auxiliar == "14957" ~	"PUIMAYEN",
      localidad_auxiliar == "14958" ~	"TRES ISLAS",
      localidad_auxiliar == "14959" ~	"ARACHANIA",
      localidad_auxiliar == "14960" ~	"PTA. RUBIA Y STA. ISABEL DE LA PEDRERA",
      localidad_auxiliar == "14961" ~	"ATLÁNTICA",
      localidad_auxiliar == "14962" ~	"PUNTA DEL DIABLO",
      localidad_auxiliar == "14963" ~	"PALMARES DE LA CORONILLA",
      localidad_auxiliar == "14964" ~	"LA RIBIERA",
      localidad_auxiliar == "14965" ~	"PUENTE VALIZAS",
      localidad_auxiliar == "14966" ~	"OCEANIA DEL POLONIO",
      localidad_auxiliar == "14967" ~	"PUEBLO NUEVO",
      localidad_auxiliar == "14968" ~	"TAJAMARES DE LA PEDRERA",
      localidad_auxiliar == "14969" ~	"SAN ANTONIO",
      localidad_auxiliar == "14970" ~	"RINCON DE LA LAGUNA",
      localidad_auxiliar == "14971" ~	"SANTA ANA DEL MAR",
      localidad_auxiliar == "14972" ~	"LOMAS DE LA PEDRERA",
      localidad_auxiliar == "15120" ~	"SALTO",
      localidad_auxiliar == "15522" ~	"BELEN",
      localidad_auxiliar == "15523" ~	"CONSTITUCION",
      localidad_auxiliar == "15721" ~	"FERNANDEZ",
      localidad_auxiliar == "15722" ~	"SAN ANTONIO",
      localidad_auxiliar == "15723" ~	"CHACRAS DE BELEN",
      localidad_auxiliar == "15724" ~	"ITAPEBÍ",
      localidad_auxiliar == "15821" ~	"ALBISU",
      localidad_auxiliar == "15822" ~	"BIASSINI",
      localidad_auxiliar == "15823" ~	"CAMPO DE TODOS",
      localidad_auxiliar == "15824" ~	"CAYETANO",
      localidad_auxiliar == "15825" ~	"CUCHILLA DE GUAVIYU",
      localidad_auxiliar == "15826" ~	"HIPÓDROMO",
      localidad_auxiliar == "15827" ~	"TERMAS DEL DAYMAN",
      localidad_auxiliar == "15828" ~	"PALOMAS",
      localidad_auxiliar == "15829" ~	"PASO DEL PARQUE DEL DAYMAN",
      localidad_auxiliar == "15830" ~	"QUINTANA",
      localidad_auxiliar == "15831" ~	"SARANDI DE ARAPEY",
      localidad_auxiliar == "15832" ~	"SAUCEDO",
      localidad_auxiliar == "15900" ~	"RURAL/PARAJE RURAL",
      localidad_auxiliar == "15921" ~	"ALCAIN",
      localidad_auxiliar == "15922" ~	"ALVEZ",
      localidad_auxiliar == "15923" ~	"SOPAS",
      localidad_auxiliar == "15924" ~	"ARENITAS BLANCAS",
      localidad_auxiliar == "15925" ~	"BALTASAR BRUM",
      localidad_auxiliar == "15926" ~	"BORDENAVE",
      localidad_auxiliar == "15927" ~	"CANCELA",
      localidad_auxiliar == "15928" ~	"CANCELA O VARESSE",
      localidad_auxiliar == "15929" ~	"CARUMBÉ",
      localidad_auxiliar == "15930" ~	"CASCO",
      localidad_auxiliar == "15931" ~	"CELESTE",
      localidad_auxiliar == "15932" ~	"CERRILLADA",
      localidad_auxiliar == "15933" ~	"CERRILLADA DE SAUCEDO",
      localidad_auxiliar == "15934" ~	"CERRO CHATO",
      localidad_auxiliar == "15935" ~	"CERROS DE VERA",
      localidad_auxiliar == "15936" ~	"EL ESPINILLAR",
      localidad_auxiliar == "15937" ~	"FARÍAS",
      localidad_auxiliar == "15938" ~	"FERREIRA",
      localidad_auxiliar == "15939" ~	"GARIBALDI",
      localidad_auxiliar == "15940" ~	"LA BOLSA",
      localidad_auxiliar == "15942" ~	"LA BOLSA 03",
      localidad_auxiliar == "15943" ~	"TERMAS DEL ARAPEY",
      localidad_auxiliar == "15944" ~	"LAS FLORES",
      localidad_auxiliar == "15945" ~	"LAURELES",
      localidad_auxiliar == "15946" ~	"LLUVERAS",
      localidad_auxiliar == "15947" ~	"MARIO RUBIO",
      localidad_auxiliar == "15948" ~	"MATAOJITO",
      localidad_auxiliar == "15949" ~	"MIGLIARO",
      localidad_auxiliar == "15950" ~	"OLIVERA",
      localidad_auxiliar == "15951" ~	"PARADA HERRERÍA",
      localidad_auxiliar == "15952" ~	"PASO DE LAS CAÑAS",
      localidad_auxiliar == "15953" ~	"PASO DEL TROPERO",
      localidad_auxiliar == "15954" ~	"PASO DEL TAPADO",
      localidad_auxiliar == "15955" ~	"PASO NUEVO DEL ARAPEY",
      localidad_auxiliar == "15956" ~	"PEPE NÚÑEZ",
      localidad_auxiliar == "15957" ~	"PASO DE LAS PIEDRAS DE ARERUNGUA",
      localidad_auxiliar == "15958" ~	"PUNTAS DE CAÑAS",
      localidad_auxiliar == "15959" ~	"PUNTAS DE VALENTIN",
      localidad_auxiliar == "15960" ~	"RAMOS",
      localidad_auxiliar == "15961" ~	"SARANDÍ",
      localidad_auxiliar == "15962" ~	"SAUCE CHICO",
      localidad_auxiliar == "15963" ~	"SOTO",
      localidad_auxiliar == "15964" ~	"TORO NEGRO",
      localidad_auxiliar == "15965" ~	"RINCON DE VALENTIN",
      localidad_auxiliar == "15966" ~	"COLONIA 18 DE JULIO",
      localidad_auxiliar == "15967" ~	"SANTA ANA",
      localidad_auxiliar == "15968" ~	"ARAPEY",
      localidad_auxiliar == "15969" ~	"PARQUE JOSE LUIS",
      localidad_auxiliar == "15970" ~	"COLONIA ITAPEBI",
      localidad_auxiliar == "15971" ~	"GUAVIYU DE ARAPEY",
      localidad_auxiliar == "15972" ~	"RUSSO",
      localidad_auxiliar == "15973" ~	"PASO CEMENTERIO",
      localidad_auxiliar == "15974" ~	"OSIMANI Y LLERENA",
      localidad_auxiliar == "15975" ~	"BARRIO FERNANDEZ",
      localidad_auxiliar == "15976" ~	"BARRIO HIPICO",
      localidad_auxiliar == "16021" ~	"CIUDAD DEL PLATA",
      localidad_auxiliar == "16022" ~	"ECILDA PAULLIER",
      localidad_auxiliar == "16220" ~	"SAN JOSE DE MAYO",
      localidad_auxiliar == "16421" ~	"LIBERTAD",
      localidad_auxiliar == "16521" ~	"DELTA DEL TIGRE Y VILLAS",
      localidad_auxiliar == "16621" ~	"RODRIGUEZ",
      localidad_auxiliar == "16721" ~	"ITUZAINGO",
      localidad_auxiliar == "16722" ~	"SANTA MÓNICA",
      localidad_auxiliar == "16793" ~	"ISMAEL CORTINAS",
      localidad_auxiliar == "16821" ~	"PUNTAS DE VALDEZ",
      localidad_auxiliar == "16822" ~	"GONZALEZ",
      localidad_auxiliar == "16823" ~	"MAL ABRIGO",
      localidad_auxiliar == "16824" ~	"PLAYA PASCUAL",
      localidad_auxiliar == "16825" ~	"18 DE JULIO (PUEBLO NUEVO)",
      localidad_auxiliar == "16826" ~	"RAFAEL PERAZA",
      localidad_auxiliar == "16827" ~	"RAIGON",
      localidad_auxiliar == "16828" ~	"TALA DE PEREIRA",
      localidad_auxiliar == "16829" ~	"SAFICI (PARQUE POSTEL)",
      localidad_auxiliar == "16900" ~	"RURAL/PARAJE RURAL",
      localidad_auxiliar == "16921" ~	"ARROYO LLANO",
      localidad_auxiliar == "16922" ~	"BELLA VISTA",
      localidad_auxiliar == "16923" ~	"JUAN SOLER",
      localidad_auxiliar == "16924" ~	"BOCA DEL CUFRE",
      localidad_auxiliar == "16925" ~	"CAPURRO",
      localidad_auxiliar == "16926" ~	"VILLA MARIA",
      localidad_auxiliar == "16927" ~	"COLONIA AMÉRICA",
      localidad_auxiliar == "16928" ~	"CUCHILLA DEL VICHADERO",
      localidad_auxiliar == "16930" ~	"ESCUDERO",
      localidad_auxiliar == "16931" ~	"FAJINA",
      localidad_auxiliar == "16932" ~	"KIYU-ORDEIG",
      localidad_auxiliar == "16933" ~	"LA BOYADA",
      localidad_auxiliar == "16934" ~	"LA BOYADA RUTA1 KM 91,2",
      localidad_auxiliar == "16935" ~	"LA CANDELARIA",
      localidad_auxiliar == "16936" ~	"LA CUCHILLA",
      localidad_auxiliar == "16937" ~	"ORDEIG",
      localidad_auxiliar == "16938" ~	"PAVÓN",
      localidad_auxiliar == "16939" ~	"CAÑADA GRANDE",
      localidad_auxiliar == "16941" ~	"LAUREL",
      localidad_auxiliar == "16942" ~	"RAPETTI",
      localidad_auxiliar == "16943" ~	"RINCÓN DE NAZARET",
      localidad_auxiliar == "16944" ~	"RINCÓN DE BUSCHENTAL",
      localidad_auxiliar == "16945" ~	"RINCON DEL PINO",
      localidad_auxiliar == "16946" ~	"SAN GREGORIO",
      localidad_auxiliar == "16947" ~	"SCAVINO",
      localidad_auxiliar == "16948" ~	"TROPAS VIEJAS",
      localidad_auxiliar == "16949" ~	"VALDEZ CHICO",
      localidad_auxiliar == "16950" ~	"MONTE GRANDE",
      localidad_auxiliar == "16951" ~	"AGUAS CORRIENTES",
      localidad_auxiliar == "16960" ~	"CERAMICAS DEL SUR",
      localidad_auxiliar == "16961" ~	"RADIAL",
      localidad_auxiliar == "16962" ~	"COLOLO TINOSA",
      localidad_auxiliar == "16963" ~	"MANGRULLO",
      localidad_auxiliar == "16964" ~	"CARRETA QUEMADA",
      localidad_auxiliar == "16965" ~	"COSTAS DE PEREIRA",
      localidad_auxiliar == "16966" ~	"COLONIA DELTA",
      localidad_auxiliar == "16967" ~	"TROPEZON",
      localidad_auxiliar == "16968" ~	"PICADA VARELA",
      localidad_auxiliar == "16970" ~	"PUNTAS DE GREGORIO",
      localidad_auxiliar == "17220" ~	"MERCEDES",
      localidad_auxiliar == "17321" ~	"DOLORES",
      localidad_auxiliar == "17521" ~	"CARDONA",
      localidad_auxiliar == "17621" ~	"JOSE ENRIQUE RODO",
      localidad_auxiliar == "17622" ~	"PALMITAS",
      localidad_auxiliar == "17623" ~	"VILLA SORIANO",
      localidad_auxiliar == "17624" ~	"PALMAR",
      localidad_auxiliar == "17721" ~	"EGAÑA",
      localidad_auxiliar == "17722" ~	"SANTA CATALINA",
      localidad_auxiliar == "17793" ~	"ISMAEL CORTINAS",
      localidad_auxiliar == "17798" ~	"AGRACIADA",
      localidad_auxiliar == "17821" ~	"CAÑADA PARAGUAYA",
      localidad_auxiliar == "17822" ~	"CASTILLOS",
      localidad_auxiliar == "17823" ~	"RISSO",
      localidad_auxiliar == "17824" ~	"SACACHISPAS",
      localidad_auxiliar == "17825" ~	"CAÑADA NIETO",
      localidad_auxiliar == "17900" ~	"RURAL/PARAJE RURAL",
      localidad_auxiliar == "17919" ~	"ISLAS",
      localidad_auxiliar == "17921" ~	"ARROYO GRANDE",
      localidad_auxiliar == "17922" ~	"BEQUELÓ RUTA 14 KM 4",
      localidad_auxiliar == "17923" ~	"BEQUELÓ RUTA 14 KM 7",
      localidad_auxiliar == "17924" ~	"BIZCOCHO",
      localidad_auxiliar == "17925" ~	"CALVO",
      localidad_auxiliar == "17926" ~	"COLONIA DÍAZ",
      localidad_auxiliar == "17927" ~	"CUCHILLA DEL PERDIDO",
      localidad_auxiliar == "17928" ~	"EL TALA",
      localidad_auxiliar == "17929" ~	"JACKSON",
      localidad_auxiliar == "17930" ~	"LA CONCORDIA",
      localidad_auxiliar == "17931" ~	"LA LOMA",
      localidad_auxiliar == "17932" ~	"OLIVERA",
      localidad_auxiliar == "17933" ~	"PAMER",
      localidad_auxiliar == "17934" ~	"RINCÓN DE COLOLÓ",
      localidad_auxiliar == "17935" ~	"SAN DIOS",
      localidad_auxiliar == "17936" ~	"SAN MARTÍN",
      localidad_auxiliar == "17937" ~	"SARANDÍ CHICO",
      localidad_auxiliar == "17938" ~	"ZANJA HONDA",
      localidad_auxiliar == "17939" ~	"PALO SOLO",
      localidad_auxiliar == "17950" ~	"CHACRAS DE DOLORES",
      localidad_auxiliar == "17951" ~	"COLONIA CONCORDIA",
      localidad_auxiliar == "17952" ~	"PERSEVERANO",
      localidad_auxiliar == "17993" ~	"LARES",
      localidad_auxiliar == "17994" ~	"PENSE",
      localidad_auxiliar == "17995" ~	"PLAYA DE LOS TREINTA Y TRES (CASA BLANCA",
      localidad_auxiliar == "18020" ~	"TACUAREMBO",
      localidad_auxiliar == "18321" ~	"PASO DE LOS TOROS",
      localidad_auxiliar == "18521" ~	"SAN GREGORIO DE POLANCO",
      localidad_auxiliar == "18695" ~	"TAMBORES",
      localidad_auxiliar == "18721" ~	"ACHAR",
      localidad_auxiliar == "18722" ~	"ANSINA",
      localidad_auxiliar == "18723" ~	"CURTINA",
      localidad_auxiliar == "18724" ~	"PASO DEL CERRO",
      localidad_auxiliar == "18725" ~	"RINCÓN DE MARTINOTE",
      localidad_auxiliar == "18822" ~	"CLARA",
      localidad_auxiliar == "18823" ~	"CLAVIJO",
      localidad_auxiliar == "18824" ~	"CUCHILLA DE CARAGUATÁ",
      localidad_auxiliar == "18825" ~	"CUCHILLA DEL OMBU",
      localidad_auxiliar == "18827" ~	"LA HILERA",
      localidad_auxiliar == "18828" ~	"LAS TOSCAS DE CARAGUATA",
      localidad_auxiliar == "18829" ~	"PASO BONILLA",
      localidad_auxiliar == "18830" ~	"PUEBLO DE ARRIBA",
      localidad_auxiliar == "18831" ~	"PUEBLO DEL BARRO",
      localidad_auxiliar == "18832" ~	"RINCÓN DE LA ALDEA",
      localidad_auxiliar == "18833" ~	"RINCON DEL BONETE",
      localidad_auxiliar == "18834" ~	"BALNEARIO IPORA",
      localidad_auxiliar == "18835" ~	"ZAPARÁ",
      localidad_auxiliar == "18853" ~	"LOS RODRÍGUEZ",
      localidad_auxiliar == "18894" ~	"PIEDRA SOLA",
      localidad_auxiliar == "18900" ~	"RURAL/PARAJE RURAL",
      localidad_auxiliar == "18921" ~	"ALDEA SAN JOAQUÍN",
      localidad_auxiliar == "18922" ~	"ATAQUES",
      localidad_auxiliar == "18923" ~	"BAÑADO DE ROCHA",
      localidad_auxiliar == "18925" ~	"CARDOZO",
      localidad_auxiliar == "18926" ~	"CARDOZO CHICO",
      localidad_auxiliar == "18927" ~	"CERRO DE LA VENTANA",
      localidad_auxiliar == "18928" ~	"CHAMBERLAIN",
      localidad_auxiliar == "18929" ~	"COLMAN",
      localidad_auxiliar == "18930" ~	"CUCHILLA DE LA PALMA",
      localidad_auxiliar == "18931" ~	"CUCHILLA DE PERALTA",
      localidad_auxiliar == "18932" ~	"HERIBERTO",
      localidad_auxiliar == "18933" ~	"LA ALDEA",
      localidad_auxiliar == "18934" ~	"LA BOLSA 01",
      localidad_auxiliar == "18936" ~	"LA HUMEDAD",
      localidad_auxiliar == "18937" ~	"LA PEDRERA",
      localidad_auxiliar == "18938" ~	"LA ROSADA",
      localidad_auxiliar == "18939" ~	"LAMBARÉ",
      localidad_auxiliar == "18940" ~	"LARRAYOS",
      localidad_auxiliar == "18941" ~	"LAS ARENAS",
      localidad_auxiliar == "18943" ~	"LA CHIRCAS",
      localidad_auxiliar == "18945" ~	"LOS FEOS",
      localidad_auxiliar == "18946" ~	"LOS FERREIRA",
      localidad_auxiliar == "18948" ~	"LOS GARCÍA",
      localidad_auxiliar == "18949" ~	"LOS GÓMEZ",
      localidad_auxiliar == "18950" ~	"LOS MAGARIÑOS",
      localidad_auxiliar == "18951" ~	"PASO DE LOS NOVILLOS",
      localidad_auxiliar == "18952" ~	"LAURELES",
      localidad_auxiliar == "18953" ~	"LOS ORTIZ",
      localidad_auxiliar == "18954" ~	"LOS ROSANOS",
      localidad_auxiliar == "18955" ~	"LOS ROSAS",
      localidad_auxiliar == "18956" ~	"LOS SEMPER",
      localidad_auxiliar == "18957" ~	"LOS VÁZQUEZ",
      localidad_auxiliar == "18958" ~	"LAURA",
      localidad_auxiliar == "18960" ~	"MATUTINA",
      localidad_auxiliar == "18961" ~	"MINUANO",
      localidad_auxiliar == "18963" ~	"MONTEVIDEO CHICO",
      localidad_auxiliar == "18964" ~	"PASO DE CEFERINO",
      localidad_auxiliar == "18965" ~	"PASO DE LAS CARRETAS",
      localidad_auxiliar == "18966" ~	"PASO DEL MEDIO",
      localidad_auxiliar == "18968" ~	"PICADA DE CUELLO",
      localidad_auxiliar == "18970" ~	"QUIEBRA YUGOS",
      localidad_auxiliar == "18971" ~	"CERRO DE PASTOREO",
      localidad_auxiliar == "18972" ~	"RINCÓN DE LA LAGUNA",
      localidad_auxiliar == "18973" ~	"RINCÓN DE GILOCA",
      localidad_auxiliar == "18975" ~	"RINCÓN DE FREITAS",
      localidad_auxiliar == "18976" ~	"RIVERA CHICO",
      localidad_auxiliar == "18978" ~	"SANTA RITA",
      localidad_auxiliar == "18979" ~	"SANTANDER",
      localidad_auxiliar == "18980" ~	"SAUCE DE BATOVI",
      localidad_auxiliar == "18981" ~	"SAUCE DE TRANQUERAS",
      localidad_auxiliar == "18982" ~	"SAUCE SOLO 02",
      localidad_auxiliar == "18984" ~	"TREINTA Y TRES Ó CAÑAS",
      localidad_auxiliar == "18985" ~	"TRES GUITARRAS",
      localidad_auxiliar == "18986" ~	"TURUPÍ",
      localidad_auxiliar == "18987" ~	"VALLE EDÉN",
      localidad_auxiliar == "18988" ~	"ZAPUCAY",
      localidad_auxiliar == "18989" ~	"CAPÓN DE LA YERBA",
      localidad_auxiliar == "18990" ~	"PUNTAS DE CINCO SAUCES",
      localidad_auxiliar == "18991" ~	"RINCON DE PEREIRA",
      localidad_auxiliar == "18992" ~	"PUNTA DE CARRETERA",
      localidad_auxiliar == "18993" ~	"CRUZ DE LOS CAMINOS",
      localidad_auxiliar == "19220" ~	"TREINTA Y TRES",
      localidad_auxiliar == "19521" ~	"SANTA CLARA DE OLIMAR",
      localidad_auxiliar == "19522" ~	"VERGARA",
      localidad_auxiliar == "19721" ~	"ARROZAL TREINTA Y TRES",
      localidad_auxiliar == "19722" ~	"GRAL. ENRIQUE MARTINEZ",
      localidad_auxiliar == "19723" ~	"VILLA SARA",
      localidad_auxiliar == "19790" ~	"CERRO CHATO",
      localidad_auxiliar == "19821" ~	"ESTACION RINCON",
      localidad_auxiliar == "19822" ~	"ISLA PATRULLA (MARIA ISABEL)",
      localidad_auxiliar == "19823" ~	"PICADA TECHERA",
      localidad_auxiliar == "19896" ~	"VALENTINES",
      localidad_auxiliar == "19900" ~	"RURAL/PARAJE RURAL",
      localidad_auxiliar == "19921" ~	"ACOSTA",
      localidad_auxiliar == "19922" ~	"POBLADO ALONZO",
      localidad_auxiliar == "19923" ~	"ARROZAL FLORENCIO BARRETO",
      localidad_auxiliar == "19924" ~	"ARROCERA RINCON",
      localidad_auxiliar == "19925" ~	"ARROZAL SANTA MARÍA",
      localidad_auxiliar == "19926" ~	"BAÑADO DE LOS OLIVERA",
      localidad_auxiliar == "19927" ~	"CAÑADA CHICA",
      localidad_auxiliar == "19928" ~	"CAÑADA DE LOS CUERVOS",
      localidad_auxiliar == "19929" ~	"CERROS DE AMARO",
      localidad_auxiliar == "19931" ~	"CIPA OLIMAR",
      localidad_auxiliar == "19932" ~	"CIPA SECADOR",
      localidad_auxiliar == "19933" ~	"COSTA DEL ARROYO MALO",
      localidad_auxiliar == "19934" ~	"CUCHILLA DE DIONISIO",
      localidad_auxiliar == "19935" ~	"EMBARQUE ARROZAL TREINTA Y TRES",
      localidad_auxiliar == "19936" ~	"JULIO MARÍA SANZ",
      localidad_auxiliar == "19937" ~	"LA CALAVERA",
      localidad_auxiliar == "19938" ~	"LA LATA",
      localidad_auxiliar == "19939" ~	"LECHIGUANA DE CORRALES",
      localidad_auxiliar == "19940" ~	"ARROCERA LOS CEIBOS",
      localidad_auxiliar == "19941" ~	"MARIA ALBINA",
      localidad_auxiliar == "19942" ~	"MENDIZABAL (EL ORO)",
      localidad_auxiliar == "19943" ~	"PASO DE PÍRIZ",
      localidad_auxiliar == "19944" ~	"PASTOR",
      localidad_auxiliar == "19945" ~	"POBLADO MEDINA",
      localidad_auxiliar == "19947" ~	"PUNTAS DEL PARAO",
      localidad_auxiliar == "19948" ~	"RINCÓN DE GADEA",
      localidad_auxiliar == "19949" ~	"RINCÓN DE LOS FRANCOS",
      localidad_auxiliar == "19950" ~	"SAN JUAN",
      localidad_auxiliar == "19951" ~	"SIERRA DEL YERBAL",
      localidad_auxiliar == "19952" ~	"SIETE CASAS",
      localidad_auxiliar == "19953" ~	"TRES BOCAS",
      localidad_auxiliar == "19954" ~	"VERDE ALTO",
      localidad_auxiliar == "19955" ~	"VILLA PASSANO",
      localidad_auxiliar == "19956" ~	"EJIDO DE TREINTA Y TRES",
      localidad_auxiliar == "19960" ~	"EL BELLACO",
      localidad_auxiliar == "19961" ~	"ARROCERA LOS TEROS",
      localidad_auxiliar == "19962" ~	"ARROCERA BONOMO",
      localidad_auxiliar == "19963" ~	"ARROCERA EL TIGRE",
      localidad_auxiliar == "19964" ~	"ARROCERA LA CATUMBERA",
      localidad_auxiliar == "19965" ~	"ARROCERA LA QUERENCIA",
      localidad_auxiliar == "19966" ~	"ARROCERA LAS PALMAS",
      localidad_auxiliar == "19967" ~	"ARROCERA MINI",
      localidad_auxiliar == "19968" ~	"ARROCERA PROCIPA",
      localidad_auxiliar == "19969" ~	"ARROCERA SAN FERNANDO",
      localidad_auxiliar == "19970" ~	"ARROCERA SANTA FE",
      localidad_auxiliar == "19971" ~	"ARROCERA ZAPATA"
    ),
    region_4 = case_when(
      region_4 == 1 ~ "Montevideo",
      region_4 == 2 ~ "Interior, localidades de 5000 habitantes o más",
      region_4 == 3 ~ "Interior, localidades de menos de 5000 habitantes",
      region_4 == 4 ~ "Zona rural"
    ),
    area = ifelse(area == 1, "Urbano", "Rural"),
    vivvo00 = ifelse(vivvo00 == 1, "Particular", "Colectiva"),
    
    ## Tenencia
    
    hogte01 = case_when(
      hogte01 == 1    ~ "Propietario",
      hogte01 == 2    ~ "Integrante de una cooperativa de vivienda",
      hogte01 == 3    ~ "Inquilino o arrendatario",
      hogte01 == 4    ~ "Usufructuario u ocupante",
      hogte01 == 7777 ~ "No corresponde",
      hogte01 == 8888 ~ "No relevado",
      hogte01 == 9898 ~ "Ignorado",
      hogte01 == 9999 ~ "No recuerda",
      hogte01 == 99   ~ "No recuerda"
    ),
    hogte02 = case_when(
      hogte02 == 1    ~ "De la vivienda y el terreno y la está pagando",
      hogte02 == 2    ~ "De la vivienda y el terreno y ya la pagó",
      hogte02 == 3    ~ "Solamente de la vivienda y la está pagando",
      hogte02 == 4    ~ "Solamente de la vivienda y ya la pagó",
      hogte02 == 7777 ~ "No corresponde",
      hogte02 == 8888 ~ "No relevado",
      hogte02 == 9898 ~ "Ignorado",
      hogte02 == 9999 ~ "No recuerda",
      hogte02 == 99   ~ "No recuerda"
    ),
    hogte03 = case_when(
      hogte03 == 1 ~	"Por relación de dependencia (trabajo)",
      hogte03 == 2 ~	"Gratuito, otorgada por el BPS",
      hogte03 == 3 ~	"Gratuito, vivienda prestada",
      hogte03 == 4 ~	"Sin permiso del propietario",
      hogte03 == 7777 ~ "No corresponde",
      hogte03 == 8888 ~ "No relevado",
      hogte03 == 9898 ~ "Ignorado",
      hogte03 == 9999 ~ "No recuerda",
      hogte03 == 99   ~ "No recuerda"
    ),
    
    ## Servicio higiénico
    
    hogsh01 = case_when(
      hogsh01 == 1    ~ "Sí, con cisterna",
      hogsh01 == 2    ~ "Sí, sin cisterna",
      hogsh01 == 3    ~ "No dispone de baño",
      hogsh01 == 7777 ~ "No corresponde",
      hogsh01 == 8888 ~ "No relevado",
      hogsh01 == 9898 ~ "Ignorado",
      hogsh01 == 9999 ~ "No recuerda",
      hogsh01 == 99   ~ "No recuerda"
    ),
    hogsh02 = case_when(
      hogsh02 == 1    ~ "De uso exclusivo del hogar",
      hogsh02 == 2    ~ "Compartido con otros hogares",
      hogsh02 == 7777 ~ "No corresponde",
      hogsh02 == 8888 ~ "No relevado",
      hogsh02 == 9898 ~ "Ignorado",
      hogsh02 == 9999 ~ "No recuerda",
      hogsh02 == 99   ~ "No recuerda"
    ),
    hogsh03 = case_when(
      hogsh03 == 1 ~ "Red general o saneamiento",
      hogsh03 == 2 ~ "Fosa séptica, pozo negro",
      hogsh03 == 3 ~ "Entubado hacia el arroyo",
      hogsh03 == 4 ~ "Superficie, hueco en el suelo u otro",
      hogsh03 == 7777 ~ "No corresponde",
      hogsh03 == 8888 ~ "No relevado",
      hogsh03 == 9898 ~ "Ignorado",
      hogsh03 == 9999 ~ "No recuerda",
      hogsh03 == 99   ~ "No recuerda"
    ),
    
    ## Servicio de cocina
    
    hogsc01 = case_when(
      hogsc01 == 1    ~	"Sí, privado de este hogar",
      hogsc01 == 2    ~	"Sí, compartido con otros hogares",
      hogsc01 == 3    ~	"No",
      hogsc01 == 7777 ~ "No corresponde",
      hogsc01 == 8888 ~ "No relevado",
      hogsc01 == 9898 ~ "Ignorado",
      hogsc01 == 9999 ~ "No recuerda",
      hogsc01 == 99   ~ "No recuerda"
    ),
    
    ## Habitaciones utilizadas con fines residenciales
    
    hoghd00 = as.character(hoghd00),
    hoghd00 = case_when(
      hoghd00  == "7777" ~ "No corresponde",
      hoghd00  == "8888" ~ "No relevado",
      hoghd00  == "9898" ~ "Ignorado",
      hoghd00  == "9999" ~ "No recuerda",
      hoghd00  == "99"   ~ "No recuerda",
      .default =  hoghd00
    ),
    hoghd01 = as.character(hoghd01),
    hoghd01 = case_when(
      hoghd01  == "7777" ~ "No corresponde",
      hoghd01  == "8888" ~ "No relevado",
      hoghd01  == "9898" ~ "Ignorado",
      hoghd01  == "9999" ~ "No recuerda",
      hoghd01  == "99"   ~ "No recuerda",
      .default =  hoghd01
    ),
    
    ## Fuente de energía para cocinar
    
    hogsc02 = case_when(
      hogsc02 == 1    ~ "Electricidad",
      hogsc02 == 2    ~ "Gas por cañería",
      hogsc02 == 3    ~ "Supergás",
      hogsc02 == 4    ~ "Queroseno",
      hogsc02 == 5    ~ "Leña o carbón",
      hogsc02 == 6    ~ "Otra fuente",
      hogsc02 == 7    ~ "Ninguna",
      hogsc02 == 7777 ~ "No corresponde",
      hogsc02 == 8888 ~ "No relevado",
      hogsc02 == 9898 ~ "Ignorado",
      hogsc02 == 9999 ~ "No recuerda",
      hogsc02 == 99   ~ "No recuerda"
    ),
    
    ## Fuente de energía para calefaccionar los ambientes
    
    hogca01 = case_when(
      hogca01 == 1 ~ "Electricidad",
      hogca01 == 2 ~ "Leña o derivados",
      hogca01 == 3 ~ "Gas por cañería",
      hogca01 == 4 ~ "Supergás",
      hogca01 == 5 ~ "Queroseno",
      hogca01 == 6 ~ "Gas oil",
      hogca01 == 7 ~ "Fuel-oil",
      hogca01 == 8 ~ "Otra fuente",
      hogca01 == 9 ~ "Ninguna",
      hogca01 == 7777 ~ "No corresponde",
      hogca01 == 8888 ~ "No relevado",
      hogca01 == 9898 ~ "Ignorado",
      hogca01 == 9999 ~ "No recuerda",
      hogca01 == 99   ~ "No recuerda"
    ),
    
    ## Disposición de residuos sólidos
    
    hogrs01 = case_when(
      hogrs01 == 1    ~ "La tiran en un contenedor o depósito",
      hogrs01 == 2    ~ "La recoge un camión o carrito de basura",
      hogrs01 == 3    ~ "La queman",
      hogrs01 == 4    ~ "La entierran",
      hogrs01 == 5    ~ "La tiran en un terreno baldío o calle",
      hogrs01 == 6    ~ "La tiran al río, arroyo o mar",
      hogrs01 == 7    ~ "Otra forma de desecharla",
      hogrs01 == 7777 ~ "No corresponde",
      hogrs01 == 8888 ~ "No relevado",
      hogrs01 == 9898 ~ "Ignorado",
      hogrs01 == 9999 ~ "No recuerda",
      hogrs01 == 99   ~ "No recuerda"
    ),
    
    ## Confort y equipamiento
    
    hogce27 = case_when(
      hogce27 == 1	  ~ "Sí",
      hogce27 == 2	  ~ "No",
      hogce27 == 7777 ~ "No corresponde",
      hogce27 == 8888 ~ "No relevado",
      hogce27 == 9898 ~ "Ignorado",
      hogce27 == 9999 ~ "No recuerda",
      hogce27 == 99   ~ "No recuerda"
    ),
    hogce26 = case_when(
      hogce26 == 1	  ~ "Sí",
      hogce26 == 2	  ~ "No",
      hogce26 == 7777 ~ "No corresponde",
      hogce26 == 8888 ~ "No relevado",
      hogce26 == 9898 ~ "Ignorado",
      hogce26 == 9999 ~ "No recuerda",
      hogce26 == 99   ~ "No recuerda"
    ),
    hogce03 = case_when(
      hogce03 == 1	  ~ "Sí",
      hogce03 == 2	  ~ "No",
      hogce03 == 7777 ~ "No corresponde",
      hogce03 == 8888 ~ "No relevado",
      hogce03 == 9898 ~ "Ignorado",
      hogce03 == 9999 ~ "No recuerda",
      hogce03 == 99   ~ "No recuerda"
    ),
    hogce04 = case_when(
      hogce04 == 1	  ~ "Sí",
      hogce04 == 2	  ~ "No",
      hogce04 == 7777 ~ "No corresponde",
      hogce04 == 8888 ~ "No relevado",
      hogce04 == 9898 ~ "Ignorado",
      hogce04 == 9999 ~ "No recuerda",
      hogce04 == 99   ~ "No recuerda"
    ),
    hogce17 = case_when(
      hogce17 == 1	  ~ "Sí",
      hogce17 == 2	  ~ "No",
      hogce17 == 7777 ~ "No corresponde",
      hogce17 == 8888 ~ "No relevado",
      hogce17 == 9898 ~ "Ignorado",
      hogce17 == 9999 ~ "No recuerda",
      hogce17 == 99   ~ "No recuerda"
    ),
    hogce17_1 = as.character(hogce17_1),
    hogce17_1 = case_when(
      hogce17_1  == "7777" ~ "No corresponde",
      hogce17_1  == "8888" ~ "No relevado",
      hogce17_1  == "9898" ~ "Ignorado",
      hogce17_1  == "9999" ~ "No recuerda",
      hogce17_1  == "99"   ~ "No recuerda",
      .default =  hogce17_1
   ),
    hogce28 = case_when(
      hogce28 == 1	  ~ "Sí",
      hogce28 == 2	  ~ "No",
      hogce28 == 7777 ~ "No corresponde",
      hogce28 == 8888 ~ "No relevado",
      hogce28 == 9898 ~ "Ignorado",
      hogce28 == 9999 ~ "No recuerda",
      hogce28 == 99   ~ "No recuerda"
    ),
    hogce11 = case_when(
      hogce11 == 1	  ~ "Sí",
      hogce11 == 2	  ~ "No",
      hogce11 == 7777 ~ "No corresponde",
      hogce11 == 8888 ~ "No relevado",
      hogce11 == 9898 ~ "Ignorado",
      hogce11 == 9999 ~ "No recuerda",
      hogce11 == 99   ~ "No recuerda"
    ),
    hogce22 = case_when(
      hogce22 == 1	  ~ "Sí",
      hogce22 == 2	  ~ "No",
      hogce22 == 7777 ~ "No corresponde",
      hogce22 == 8888 ~ "No relevado",
      hogce22 == 9898 ~ "Ignorado",
      hogce22 == 9999 ~ "No recuerda",
      hogce22 == 99   ~ "No recuerda"
    ),
    hogce12_1 = case_when(
      hogce12_1 == 1	  ~ "Sí",
      hogce12_1 == 2	  ~ "No",
      hogce12_1 == 7777 ~ "No corresponde",
      hogce12_1 == 8888 ~ "No relevado",
      hogce12_1 == 9898 ~ "Ignorado",
      hogce12_1 == 9999 ~ "No recuerda",
      hogce12_1 == 99   ~ "No recuerda"
    ),
   hogce13 = as.character(hogce13),
   hogce13 = case_when(
     hogce13  == "7777" ~ "No corresponde",
     hogce13  == "8888" ~ "No relevado",
     hogce13  == "9898" ~ "Ignorado",
     hogce13  == "9999" ~ "No recuerda",
     hogce13  == "99"   ~ "No recuerda",
     .default =  hogce13
   ),
   hogma01_1 = case_when(
     hogma01_1  == 1    ~ "Sí",
     hogma01_1  == 2    ~ "No",
     hogma01_1  == 7777 ~ "No corresponde",
     hogma01_1  == 8888 ~ "No relevado",
     hogma01_1  == 9898 ~ "Ignorado",
     hogma01_1  == 9999 ~ "No recuerda",
     hogma01_1  == 99   ~ "No recuerda"
   ),
   hogma01_1_1 = as.character(hogma01_1_1),
   hogma01_1_1 = case_when(
     hogma01_1_1  == "7777" ~ "No corresponde",
     hogma01_1_1  == "8888" ~ "No relevado",
     hogma01_1_1  == "9898" ~ "Ignorado",
     hogma01_1_1  == "9999" ~ "No recuerda",
     hogma01_1_1  == "99"   ~ "No recuerda",
     .default     =  hogma01_1_1
   ),
   hogma01_2 = case_when(
     hogma01_2  == 1    ~ "Sí",
     hogma01_2  == 2    ~ "No",
     hogma01_2  == 7777 ~ "No corresponde",
     hogma01_2  == 8888 ~ "No relevado",
     hogma01_2  == 9898 ~ "Ignorado",
     hogma01_2  == 9999 ~ "No recuerda",
     hogma01_2  == 99   ~ "No recuerda"
   ),
   hogma01_2_1 = as.character(hogma01_2_1),
   hogma01_2_1 = case_when(
     hogma01_2_1  == "7777" ~ "No corresponde",
     hogma01_2_1  == "8888" ~ "No relevado",
     hogma01_2_1  == "9898" ~ "Ignorado",
     hogma01_2_1  == "9999" ~ "No recuerda",
     hogma01_2_1  == "99"   ~ "No recuerda",
     .default     =  hogma01_2_1
   )) %>% 
  select(- x1, - departamento_auxiliar)

write_rds(hogar, here("data", "02_final", "hogar.rds"))
write_dta(hogar, here("data", "02_final", "hogar.dta"))
